% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\name{animalNameFromSessionName}
\alias{animalNameFromSessionName}
\title{Get animal name from session name}
\usage{
animalNameFromSessionName(sessionName = NULL)
}
\arguments{
\item{sessionName}{Character vector with the session name}
}
\value{
Character vector with animal name
}
\description{
Assumes the session name is in the format name-date-rest
}
