% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{autocorrelationDoughnut}
\alias{autocorrelationDoughnut}
\alias{autocorrelationDoughnut,ANY,ANY-method}
\alias{autocorrelationDoughnut,SpatialProperties2d-method}
\title{Calculate the spatial autocorrelation of the firing rate maps of neurons and get the region use for grid score calculation
Autocorrelation is performed on the firing rate maps of the SpatialProperties2d object.
Fields are detected as when one wants to calculate grid scores. 
Then the region with the 6 fields surrounding the center is kept
Usefull to test the different steps of the calculation of the grid score.}
\usage{
autocorrelationDoughnut(sp)

\S4method{autocorrelationDoughnut}{SpatialProperties2d}(sp)
}
\arguments{
\item{sp}{SpatialProperties2d object}
}
\value{
SpatialProperties2d object with the spatial autocorrelation in slot autosDoughnut
}
\description{
Calculate the spatial autocorrelation of the firing rate maps of neurons and get the region use for grid score calculation
Autocorrelation is performed on the firing rate maps of the SpatialProperties2d object.
Fields are detected as when one wants to calculate grid scores. 
Then the region with the 6 fields surrounding the center is kept
Usefull to test the different steps of the calculation of the grid score.
}

