% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{autocorrelationDoughnutRotate}
\alias{autocorrelationDoughnutRotate}
\alias{autocorrelationDoughnutRotate,SpatialProperties2d-method}
\alias{autocorrelationDoughnutRotate,ANY,ANY-method}
\title{Get the rotated region of the spatial autocorrelation of the firing rate maps when calculating grid scores
Autocorrelation is performed on the firing rate maps of the SpatialProperties2d object.
Fields are detected as when one wants to calculate grid scores. 
Then the region with the 6 fields surrounding the center is kept
The rotated copy of the region are stored in slot autosDoughnutRotate
Useful to test the different steps of the calculation of the grid score.}
\usage{
autocorrelationDoughnutRotate(sp)

\S4method{autocorrelationDoughnutRotate}{SpatialProperties2d}(sp)
}
\arguments{
\item{sp}{SpatialProperties2d object}
}
\value{
SpatialProperties2d object with the spatial autocorrelation in slot autosDoughnutRotate
}
\description{
Get the rotated region of the spatial autocorrelation of the firing rate maps when calculating grid scores
Autocorrelation is performed on the firing rate maps of the SpatialProperties2d object.
Fields are detected as when one wants to calculate grid scores. 
Then the region with the 6 fields surrounding the center is kept
The rotated copy of the region are stored in slot autosDoughnutRotate
Useful to test the different steps of the calculation of the grid score.
}
