% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{autocorrelationNoFields}
\alias{autocorrelationNoFields}
\alias{autocorrelationNoFields,ANY,ANY-method}
\alias{autocorrelationNoFields,SpatialProperties2d-method}
\title{Calculate the spatial autocorrelation of the firing rate maps of neurons and removed the detected fields}
\usage{
autocorrelationNoFields(sp)

\S4method{autocorrelationNoFields}{SpatialProperties2d}(sp)
}
\arguments{
\item{sp}{SpatialProperties2d object}
}
\value{
SpatialProperties2d object with the spatial autocorrelation in slot autosDetect
}
\description{
The autocorrelation is performed on the firing rate maps of the SpatialProperties2d object.
Fields are detected as when one wants to calculate grid scores
Useful to test the different steps of the calculation of the grid score.
}

