% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalProcessing.R
\name{bandPassFilter}
\alias{bandPassFilter}
\title{Apply a band-pass filter to a signal}
\usage{
bandPassFilter(data, samplingRate, minPassHz, maxPassHz)
}
\arguments{
\item{data}{Numeric vector containing the data to filter}

\item{samplingRate}{Sampling rate of the data in Hz}

\item{minPassHz}{Minimal frequency in Hz that is kept}

\item{maxPassHz}{Maximal frequency in Hz that is kept}
}
\value{
Filtered copy of the argument data
}
\description{
Use a Butterworth filter.
}

