% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{borderDetection}
\alias{borderDetection}
\alias{borderDetection,ANY,ANY-method}
\alias{borderDetection,SpatialProperties2d-method}
\title{Detect the border pixels in an environment and return
a matrix with border pixels identified.}
\usage{
borderDetection(sp, ...)

\S4method{borderDetection}{SpatialProperties2d}(sp, border = "rectangular")
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{border}{Set to rectangular or circular depending of the type of environment
Default value is rectangular}
}
\value{
A matrix with the border pixels set to non negative positive values
}
\description{
Call firingRateMap2d() to construct the map first before calling this function
}

