% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{borderScore}
\alias{borderScore}
\alias{borderScore,ANY,ANY-method}
\alias{borderScore,SpatialProperties2d-method}
\title{Calculate border scores from the firing maps}
\usage{
borderScore(sp, border = "rectangular")

\S4method{borderScore}{SpatialProperties2d}(sp, border = "rectangular")
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{border}{Set to rectangular or circular depending on the shape of environment}
}
\value{
SpatialProperties2d object with the border scores in the slot borderScore
}
\description{
Calculate border scores from the firing maps
}

