% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsGeneric.R
\name{boxplotTwoFactors}
\alias{boxplotTwoFactors}
\title{Plot a boxplot with 2 factors}
\usage{
boxplotTwoFactors(data, dv = "vd", iv1 = "vi1", iv2 = "vi2",
  colors = c("red", "gray", "red", "gray"), ylim = c(0, 1), ylab = "",
  outma = c(0.5, 1, 0.5, 1), margin = c(1, 2, 1, 1), mgp = c(1, 0.2, 0),
  ...)
}
\arguments{
\item{data}{data.frame with the data}

\item{dv}{Dependent variable. The one you want on the y-axis}

\item{iv1}{First independent variable}

\item{iv2}{Second independent variable}

\item{ylim}{Limit of y axis}

\item{ylab}{Label of y axis}

\item{outma}{outma for setting par}

\item{margin}{margin for setting par}

\item{mgp}{mgp for setting par}

\item{color}{Color of the bars of the boxplot}
}
\description{
Just a wrapper around boxplot function.
To set which level is plot first, consider using the function factor and set the argument levels.
}

