% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellGroup.R
\docType{methods}
\name{brainRegionFromCluNo}
\alias{brainRegionFromCluNo}
\alias{brainRegionFromCluNo,ANY,ANY-method}
\alias{brainRegionFromCluNo,CellGroup-method}
\title{Get the brain region from a list of cluNo}
\usage{
brainRegionFromCluNo(cg, cluNo)

\S4method{brainRegionFromCluNo}{CellGroup}(cg, cluNo)
}
\arguments{
\item{cg}{A CellGroup object}

\item{cluNo}{Numeric containing the cluNo of the cluster for which you want the brain region.}
}
\value{
Character vector with be name of the brain region for each cluNo
}
\description{
Get the brain region from a list of cluNo
}

