% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellGroup.R
\docType{class}
\name{CellGroup-class}
\alias{CellGroup}
\alias{CellGroup-class}
\title{An S4 class representing a group of cells}
\description{
This class gets you the tetrode id and brain region associated with each cell of a recording session. 
It also get the cluster id of the cell on its respective tetrode.
}
\section{Slots}{

\describe{
\item{\code{session}}{A character vector containing the names of the recording session.}

\item{\code{path}}{The directory in which the files of the session are located.}

\item{\code{fileBase}}{Is the path and session}

\item{\code{nTetrodes}}{Number of tetrodes}

\item{\code{nCells}}{Number of cells}

\item{\code{id}}{Cell id}

\item{\code{animal}}{Name of the animal}

\item{\code{clu}}{Cluster id of the cells}

\item{\code{tetrode}}{tetrode number of the cells}

\item{\code{tetrodeId}}{id of the tetrode for each cell}

\item{\code{cluToTetrode}}{clu id of the cell on its respective tetrode (used for clustering)}

\item{\code{brainRegion}}{region in which the cell was recorded}
}}
\examples{
df<-new("CellGroup")

}

