% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{centerOfMass}
\alias{centerOfMass}
\title{Calculate the center of mass of a matrix, numeric or integer}
\usage{
centerOfMass(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Numeric of length 2 with the x and y coordinate of the center fo mass
}
\description{
The values returned are in indices with first bin being 1 and last being length(x)
}
