% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whdFromPositrack.R
\name{checkIntegrityPositrackData}
\alias{checkIntegrityPositrackData}
\title{Check the integrity of positrack data read from positrack file.}
\usage{
checkIntegrityPositrackData(posi, maxDelayCapProc = 1000,
  maxInterCapDelay = 1000, maxInterProcDelay = 1000,
  maxProcDuration = 1000)
}
\arguments{
\item{posi}{Data frame containing the positrack file read via read.table(fn,header=T)}

\item{maxDelayCapProc}{Maximum allowed delay between frame capture and processing}

\item{maxInterCapDelay}{Maximum allowed delay between the time stamps of frame capture}

\item{maxInterProcDelay}{Maximum allowed delay between the time stamps of frame processing}

\item{maxProcDuration}{Maximum allowed processing of frame duration}
}
\value{
Return 0 if all is ok and positive number if something is wrong
}
\description{
Check the integrity of positrack data read from positrack file.
}
