% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\name{cluNoFromCluId}
\alias{cluNoFromCluId}
\title{Get cluNo from cluId}
\usage{
cluNoFromCluId(cluId = NULL)
}
\arguments{
\item{cluId}{Character vector with the cluIds}
}
\value{
Numeric vectors with cluNo
}
\description{
Assumes the cluId is in the format name-date-rest_cluNo
}

