% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clusterIsolationCheck}
\alias{clusterIsolationCheck}
\title{Check if clusters of a recording session are well isolated from one another}
\usage{
clusterIsolationCheck(rs)
}
\arguments{
\item{rs}{RecSession}
}
\value{
A list including isolation distance and refractory ratio of all clusters. The check element is a
logical indicating which clusters are likely to be poorly isolated.
}
\description{
This function generate a data.frame with the isolation distance and refractory period ratio for all clusters.
It is returned within a list to conform to runOnSessionList policy on returned object.
}

