% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsSpatial.R
\name{colorScalePlot}
\alias{colorScalePlot}
\title{Plot a color scale for the firing rate maps}
\usage{
colorScalePlot(outma = c(0.2, 0.2, 0.2, 0.2), margin = c(0, 1, 0, 1),
  cex.labels = 0.5)
}
\arguments{
\item{outma}{Outer margins of the figure}

\item{margin}{Inner margins of the figure}

\item{cex.labels}{Size of the font use for 0 and Max}
}
\description{
Plot a color sclae that can be used beside the plots obtained with firingRateMapPlot()
}

