% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{containsElectrodeLocation}
\alias{containsElectrodeLocation}
\alias{containsElectrodeLocation,RecSession-method}
\alias{containsElectrodeLocation,ANY,ANY-method}
\title{Check if the session had an electrode in a particular brain area}
\usage{
containsElectrodeLocation(rs, location = "")

\S4method{containsElectrodeLocation}{RecSession}(rs, location = "")
}
\arguments{
\item{rs}{A RecSession object}

\item{location}{A brain area of interest}
}
\value{
TRUE or FALSE
}
\description{
This will check whether the value of location is in the electrodeLocation vector.
}
