% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{containsEnvironment}
\alias{containsEnvironment}
\alias{containsEnvironment,RecSession-method}
\alias{containsEnvironment,ANY,ANY-method}
\title{Check if the session had a trial in a given environment}
\usage{
containsEnvironment(rs, environment = "")

\S4method{containsEnvironment}{RecSession}(rs, environment = "")
}
\arguments{
\item{rs}{A RecSession object}

\item{environment}{The name of an environment}
}
\value{
TRUE or FALSE
}
\description{
This will check whether the value of environment is in the env vector.
}
