% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{containsStimulation}
\alias{containsStimulation}
\alias{containsStimulation,ANY,ANY-method}
\alias{containsStimulation,RecSession-method}
\title{Check if the session had a trial in a given stimulation type}
\usage{
containsStimulation(rs, stimulation = "")

\S4method{containsStimulation}{RecSession}(rs, stimulation = "")
}
\arguments{
\item{rs}{A RecSession object}

\item{stimulation}{The name of a stimulation}
}
\value{
TRUE or FALSE
}
\description{
This will check whether the value of stimulation is in the stim vector.
}

