% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{methods}
\name{copyDatabase}
\alias{copyDatabase}
\alias{copyDatabase,ElectroProject-method}
\alias{copyDatabase,ANY,ANY-method}
\title{Create a copy of the data from an experiment}
\usage{
copyDatabase(ep, sessionList, copyType = "medium", extensions, destination,
  compress = FALSE)

\S4method{copyDatabase}{ElectroProject}(ep, sessionList, copyType = "medium",
  extensions, destination, compress = FALSE)
}
\arguments{
\item{ep}{ElectroProject object}

\item{sessionList}{List of RecSession objects that will be included in the copy}

\item{copyType}{Type of copy. Only "medium" is supported at the moment.}

\item{extensions}{Character vector with the list of file types you want to export}

\item{destination}{Path of the directory in which the copy is made}

\item{compress}{Logical indicating whether to call tar to make a tar.gzip from the copy}
}
\description{
There is an argument to specify the type of backup you want.
This can be use to backup different files depending on
what needs to be copied.
}
\details{
List of files exported if copyType is set to medium
- .res, .clu, .par, .desen, .px_per_cm, .resofs, .desel
.res_samples_per_whl_sample .sampling_rate_dat, .whl .whd
Also includes the tetrode specific .clu. and .res. files
}
