% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{cor.diff}
\alias{cor.diff}
\title{Perform r-to-Z transform to get significance of difference between two correlation coefficients}
\usage{
cor.diff(r1, n1, r2, n2)
}
\arguments{
\item{r1}{correlation coefficient of the first correlation}

\item{n1}{number of observations in the first correlation}

\item{r2}{correlation coefficient of the second correlation}

\item{n2}{number of observations in the second correlation}
}
\description{
Perform r-to-Z transform to get significance of difference between two correlation coefficients
}

