% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{createSpkFile}
\alias{createSpkFile}
\title{Create the spk file for each tetrode
   
 Use the fil file for each channel}
\usage{
createSpkFile(traces, rs, res, window = 32, tetrodeNumber)
}
\arguments{
\item{traces}{Matrix with the electrophysiological traces}

\item{rs}{RecSession}

\item{res}{Numeric with the time stamps of the spikes in sample value}

\item{window}{Number of ms that will be shown}

\item{tetrodeNumber}{Tetrode number used in the file name}
}
\description{
Create the spk file for each tetrode
   
 Use the fil file for each channel
}

