% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatFiles.R
\docType{class}
\name{DatFiles-class}
\alias{DatFiles-class}
\alias{DatFiles}
\title{A class representing one or more .dat files}
\description{
This class is used to read data from one or several .dat files.
Because the .dat files are binary files, you can't read them with the usual read functions within R.
If several dat files are set in an object, they are treated as though they are a single file.
}
\details{
Dat files have a very simple organisation. Each data point has 2 bytes. 
There is no header. The data from sample 0 appear first, then sample 1, etc.
If c is for channel and s is for sample, a file with 3 channel would look like this
s0c0, s0c1, s0c2, s1c0, s1c1, s1c2, s2c0, s2c1, s2c2, etc.
}
\section{Slots}{

\describe{
\item{\code{fileNames}}{A character vector containing the names of the files}

\item{\code{path}}{The directory in which the files are located}

\item{\code{samples}}{Numeric vector with the number of samples in each file.}

\item{\code{size}}{Numeric vector with the size of the file in bytes}

\item{\code{nChannels}}{Number of recorded channels in the dat files}
}}

