% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatFiles.R
\docType{methods}
\name{datFilesGetGroupChannels}
\alias{datFilesGetGroupChannels}
\alias{datFilesGetGroupChannels,ANY,ANY-method}
\alias{datFilesGetGroupChannels,DatFiles-method}
\title{Function to read from several channels from a group of dat files}
\usage{
datFilesGetGroupChannels(df, channels, firstSample, lastSample)

\S4method{datFilesGetGroupChannels}{DatFiles}(df, channels, firstSample,
  lastSample)
}
\arguments{
\item{df}{A DatFiles object}

\item{channels}{Numeric indicating which channels are needed. The first channelNo in the file is 0.}

\item{firstSample}{Numeric vector containing the first sample to retrieve. 
You can give several if you want to get a series of data chunks.
First sample in the file is 0}

\item{lastSample}{Numeric vector containing the last sample to retrieve. You can give several if you want to get a series of data chunks.
Index stats at 0}
}
\value{
A matrix containing the data, one channel per column.
}
\description{
This function returns the data from firstSample to lastSample for a group of channels. 
If firstSample and lastSample have a length greater than 1, 
the data chunks will be pasted one after the other in the data returned by the function.
}

