% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatFiles.R
\docType{methods}
\name{datFilesGetOneChannel}
\alias{datFilesGetOneChannel}
\alias{datFilesGetOneChannel,ANY,ANY-method}
\alias{datFilesGetOneChannel,DatFiles-method}
\title{Function to read one channel from a group of dat files}
\usage{
datFilesGetOneChannel(df, channelNo, firstSample, lastSample)

\S4method{datFilesGetOneChannel}{DatFiles}(df, channelNo, firstSample,
  lastSample)
}
\arguments{
\item{df}{A DatFiles object}

\item{channelNo}{The channel to retrieve the data from. The first channelNo in the file is 0.}

\item{firstSample}{Numeric vector containing the first sample to retrieve. 
You can give several if you want to get a series of data chunks.
First sample in the file is 0}

\item{lastSample}{Numeric vector containing the last sample to retrieve. You can give several if you want to get a series of data chunks.
Index stats at 0}
}
\value{
A integer vector containing the data read from the dat files.
}
\description{
This function returns the data from firstSample to lastSample. 
If firstSample and lastSample have a length greater than 1, 
the data chunks will be pasted one after the other in the data returned by the function.
}

