% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatFiles.R
\docType{methods}
\name{datFilesSamples}
\alias{datFilesSamples}
\alias{datFilesSamples,ANY,ANY-method}
\alias{datFilesSamples,DatFiles-method}
\title{Get the number of samples per file and file size}
\usage{
datFilesSamples(df)

\S4method{datFilesSamples}{DatFiles}(df)
}
\arguments{
\item{df}{A DatFiles object}
}
\value{
A DatFiles object with the samples and size set.
}
\description{
Get the number of samples per file and file size
}

