% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatFiles.R
\docType{methods}
\name{datFilesSet}
\alias{datFilesSet}
\alias{datFilesSet,DatFiles-method}
\alias{datFilesSet,ANY,ANY-method}
\title{Function to set the parameters of a DatFile object}
\usage{
datFilesSet(df, fileNames, path, nChannels)

\S4method{datFilesSet}{DatFiles}(df, fileNames = "", path = "",
  nChannels = 0)
}
\arguments{
\item{df}{A DatFiles object}

\item{fileNames}{A character vector containing the names of the .dat files}

\item{path}{A directory where the files are located. If empty the working directory is used.}

\item{nChannels}{The number of channels in the dat files}
}
\value{
A DatFiles object with the values set.
}
\description{
Function to set the parameters of a DatFile object
}
