% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{deleteCluster}
\alias{deleteCluster}
\title{Delete a specific cluster in a recording session}
\usage{
deleteCluster(rs, clu)
}
\arguments{
\item{rs}{RecSession}

\item{clu}{Cluster number. This is the session wide cluster number.
Not the tetrode specific number}
}
\description{
This will delete the cluster in the tetrode specific clu file
The deleted cluster is assigned a value of 1 in the clu file.
Don't forget to recreate the new session wide clu and res file with
the function mergeTetrodeSpecificResCluFiles()
}
