% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttl.R
\name{detectDowns}
\alias{detectDowns}
\title{Detect the negative deflection of a ttl pulse in a vector}
\usage{
detectDowns(x, threshold = 10000)
}
\arguments{
\item{x}{numeric vector containing the data}

\item{threshold}{amplitude of the negative deflections to be considered}
}
\value{
indices at which the negative deflections were detected in \code{x}
}
\description{
Detect the negative deflection of a ttl pulse in a vector
}
\examples{
 detectDowns(x=c(rep(0,10),rep(20000,10)))
}

