% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{detectFiringFields}
\alias{detectFiringFields}
\alias{detectFiringFields,ANY,ANY-method}
\alias{detectFiringFields,SpatialProperties2d-method}
\title{Detect firing fields in the firing rate maps of a SpatialProperties2d object}
\usage{
detectFiringFields(sp, minAreaCm2, rateThresholds)

\S4method{detectFiringFields}{SpatialProperties2d}(sp, minAreaCm2 = 20,
  rateThresholds)
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{minAreaCm2}{Minimal area of the field in cm squared}

\item{rateThresholds}{Numeric vector containing the firing rate threshold to be part of a firing field.
Should have the same length than the number of cells in the SpatialProperties2d object}
}
\value{
matrix with the field information (clu, xcom, ycom, peak.rate, area in cm^2)
}
\description{
Detect firing fields in the firing rate maps of a SpatialProperties2d object
}

