% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{detectSpikesFromTrace}
\alias{detectSpikesFromTrace}
\title{Detect spikes in a signal}
\usage{
detectSpikesFromTrace(data, samplingRate, powerWindowSizeMs, powerWindowSlideMs,
  SDThreshold)
}
\arguments{
\item{data}{Vector containing the spikes and some noise}

\item{samplingRate}{Sampling rate of the trace}

\item{powerWindowSizeMs}{Window size when calculating power (root mean square)}

\item{powerWindowSlideMs}{Shift of the window in ms between estimation of power}

\item{SDThreshold}{Power threshold for spike detection.}
}
\value{
list containing rms, rmsT, rmsSD, rmsMean, rmsThreshold, spikeTime and spikePower
}
\description{
If there are low components in the signal, they should be filtered out.
}

