% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttl.R
\name{detectUps}
\alias{detectUps}
\title{Detect the positive deflection of a ttl pulse in a vector}
\usage{
detectUps(x, threshold = 10000)
}
\arguments{
\item{x}{numeric vector containing the data}

\item{threshold}{amplitude of the positive deflections to be considered}
}
\value{
indices at which the positive deflections were detected in \code{x}
}
\description{
Detect the positive deflection of a ttl pulse in a vector
}
\examples{
 detectUps(x=c(rep(0,10),rep(20000,10)))
}

