% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{directionFilter}
\alias{directionFilter}
\alias{directionFilter,ANY,ANY-method}
\alias{directionFilter,Positrack-method}
\title{Set position data to NA if animal's direction was not the one given in arguments}
\usage{
directionFilter(pt, direction)

\S4method{directionFilter}{Positrack}(pt, direction = 0)
}
\arguments{
\item{pt}{Positrack object}

\item{direction}{Value of 1 or 0 refereing to direction}
}
\value{
Positrack object with only valid value in the specific direction.
}
\description{
The slots x, y, hd, speed and lin are affected.
Please note that this is not head direction but rather the direction of movement in 1D environment
}

