% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{class}
\name{ElectroProject-class}
\alias{ElectroProject}
\alias{ElectroProject-class}
\title{A S4 class to represent an electrophysiological project containing several recording sessions.}
\description{
An ElectroProject object contains a group of RecSession objects. It can be used to run functions on several recording sessions.
}
\details{
It assumes that your data are organized in a specific way on your hard drive.
There should be a top directory containing one directory for each subject of the experiment.
Within each subject directory, there should be a directory for each recording session with this subject.
By convention, the names of the session directories contain at least one hyphen.
There should be no other directories with a hyphen in their names.

This object can be used to get an overview of the progress during the data acquisition period.

Two useful methods are getSessionList and runOnSessionList
}
\section{Slots}{

\describe{
\item{\code{directory}}{Top directory of the project}

\item{\code{resultsDirectory}}{Directory where results will be saved by default}

\item{\code{nSessions}}{Number of recording sessions in the project}

\item{\code{sessionNameList}}{Name of the recording sessions}

\item{\code{sessionPathList}}{Path to all the recording sessions}

\item{\code{sessionList}}{List of RecSession objects.}
}}

