% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{fileExists}
\alias{fileExists}
\alias{fileExists,ANY,ANY-method}
\alias{fileExists,RecSession-method}
\title{Check if the session directory contains a file ending with the value of the argument extension}
\usage{
fileExists(rs, extension = "")

\S4method{fileExists}{RecSession}(rs, extension = "")
}
\arguments{
\item{rs}{A RecSession object}

\item{extension}{The extension of the file you are looking for.}
}
\value{
TRUE or FALSE
}
\description{
By default test whether paste(rs@fileBase,extension,sep=".") exists
}

