% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties1d.R
\docType{methods}
\name{firingRateHisto}
\alias{firingRateHisto}
\alias{firingRateHisto,SpatialProperties1d-method}
\alias{firingRateHisto,ANY,ANY-method}
\title{Calculate the firing rate histograms of neurons using a SpikeTrain and Positrack objects}
\usage{
firingRateHisto(sp1, st, pt)

\S4method{firingRateHisto}{SpatialProperties1d}(sp1, st, pt)
}
\arguments{
\item{sp1}{SpatialProperties1d object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack object with valid linearized position data}
}
\value{
SpatialProperties1d object with the firing rate histograms
}
\description{
Calculate the firing rate histograms of neurons using a SpikeTrain and Positrack objects
}
