% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{firingRateMap2d}
\alias{firingRateMap2d}
\alias{firingRateMap2d,ANY,ANY-method}
\alias{firingRateMap2d,SpatialProperties2d-method}
\title{Calculate the firing rate maps of neurons using a SpikeTrain and Positrack objects}
\usage{
firingRateMap2d(sp, st, pt, nRowMap = NA, nColMap = NA)

\S4method{firingRateMap2d}{SpatialProperties2d}(sp, st, pt, nRowMap = NA,
  nColMap = NA)
}
\arguments{
\item{sp}{SpatialProperties1d object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack object}

\item{nRowMap}{Numeric indicating the number of rows in the map. By default the maps has the smallest size possible given the
position data in the Positrack object. Use this argument to have maps of a fixed size. Needs to be as large as the minimal size of the map}

\item{nColMap}{Numeric indicating the number of columns in the map. By default the maps has the smallest size possible given the
position data in the Positrack object. Use this argument to have maps of a fixed size. Needs to be as large as the minimal size of the map}
}
\value{
SpatialProperties2d object with the firing rate maps
}
\description{
The occupancy map and the firing rate maps are smoothed with a Gaussian kernel
The amount of smoothing is determined by slots smoothOccupancySd and smoothRateMapSd of sp
You can force the function to create maps with arbitrary x and y sizes as long as the map fits into this arbitrary size
}

