% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{firingRateMapAutosPlot}
\alias{firingRateMapAutosPlot}
\title{Plot a several spatial autocorrelation maps on the same page}
\usage{
firingRateMapAutosPlot(maps, names, fn = "page.full.plot.pdf")
}
\arguments{
\item{maps}{A 3d array containing maps (x,y,clu)}

\item{names}{A character vector containing the name of each map in the array}

\item{fn}{Character vector containing the file name for the plot}
}
\description{
This is not currently being developed.
}

