% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{firingRateMapCenterOfMass}
\alias{firingRateMapCenterOfMass}
\alias{firingRateMapCenterOfMass,ANY,ANY-method}
\alias{firingRateMapCenterOfMass,SpatialProperties2d-method}
\title{Calculate the center of mass of the firing rate maps in a SpatialProperties2d object}
\usage{
firingRateMapCenterOfMass(sp)

\S4method{firingRateMapCenterOfMass}{SpatialProperties2d}(sp)
}
\arguments{
\item{sp1}{SpatialProperties2d object}
}
\value{
matrix containing the center of mass of the firing rate maps
}
\description{
Calculate the center of mass of the firing rate maps in a SpatialProperties2d object
}

