% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{firingRateMapCorrelation}
\alias{firingRateMapCorrelation}
\alias{firingRateMapCorrelation,SpatialProperties2d-method}
\alias{firingRateMapCorrelation,ANY,ANY-method}
\title{Calculate a Pearson correlation coefficients between the firing rate maps of two SpatialProperties2d objects}
\usage{
firingRateMapCorrelation(sp1, sp2)

\S4method{firingRateMapCorrelation}{SpatialProperties2d}(sp1, sp2)
}
\arguments{
\item{sp1}{SpatialProperties2d object}

\item{sp2}{SpatialProperties2d object}
}
\value{
Numeric containing the Pearson correlation coefficients
}
\description{
Used to estimate map stability between the firing rate maps in two conditions.
}
\details{
The maps of 2 SpatialProperties2d are compared. 
If the maps are made with different Positrack objects, make sure that the sp@reduceSize is set to FALSE.
Also make sure that the maps of the 2 SpatialProperties2d objects have the same dimensions by using the 
arguments nRowMap and nColMap when calling firingRateMap2d().
}
