% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{firingRateMapsPlot}
\alias{firingRateMapsPlot}
\title{Plot a several firing rate maps on the same page}
\usage{
firingRateMapsPlot(maps, names, fn = "page.full.plot.pdf")
}
\arguments{
\item{maps}{A 3d array containing maps (x,y,clu)}

\item{names}{A character vector containing the name of each map in the array}

\item{fn}{Character vector containing the file name for the plot}
}
\description{
Plot 2-dimensional representation of firing rate for several maps
This is not currently being developed.
}

