% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{methods}
\name{getClusteredSessionList}
\alias{getClusteredSessionList}
\alias{getClusteredSessionList,ANY,ANY-method}
\alias{getClusteredSessionList,ElectroProject-method}
\title{Return a list of clustered RecSession objects}
\usage{
getClusteredSessionList(ep)

\S4method{getClusteredSessionList}{ElectroProject}(ep)
}
\arguments{
\item{ep}{ElectroProject object}
}
\value{
list of clustered RecSession objects
}
\description{
Return a list of clustered RecSession objects
}

