% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties1d.R
\docType{methods}
\name{getHistoStats}
\alias{getHistoStats}
\alias{getHistoStats,SpatialProperties1d-method}
\alias{getHistoStats,ANY,ANY-method}
\title{Calculate spatial statistics of the firing rate histograms of neurons}
\usage{
getHistoStats(sp1, st, pt)

\S4method{getHistoStats}{SpatialProperties1d}(sp1, st, pt)
}
\arguments{
\item{sp1}{SpatialProperties1d object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack object}
}
\value{
SpatialProperties1d object with the stats in the following slots: peakRate, infoScore, sparsity
}
\description{
Calculate spatial statistics of the firing rate histograms of neurons
}
