% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties1d.R
\docType{methods}
\name{getHistoStatsShuffle}
\alias{getHistoStatsShuffle}
\alias{getHistoStatsShuffle,ANY,ANY-method}
\alias{getHistoStatsShuffle,SpatialProperties1d-method}
\title{Get random values for statistics of the firing rate histograms of neurons.}
\usage{
getHistoStatsShuffle(sp1, st, pt)

\S4method{getHistoStatsShuffle}{SpatialProperties1d}(sp1, st, pt)
}
\arguments{
\item{sp1}{SpatialProperties1d object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack object}
}
\value{
SpatialProperties1d object with the stats in the following slots: peakRateShuffle, infoScoreShuffle, sparsityShuffle
}
\description{
The random values are obtained by shifting the position data. See shiftLinRandom.
The number of shufflings is nShufflings of the SpatialProperties1d object.
}

