% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{getIntervalsAtDirection}
\alias{getIntervalsAtDirection}
\alias{getIntervalsAtDirection,Positrack-method}
\alias{getIntervalsAtDirection,ANY,ANY-method}
\title{Get time intervals at which the animal ran in a given direction (0 or 1)}
\usage{
getIntervalsAtDirection(pt, direction)

\S4method{getIntervalsAtDirection}{Positrack}(pt, direction = 0)
}
\arguments{
\item{pt}{Positrack object}

\item{direction}{0 or 1 indicating the direction of movment}
}
\value{
matrix with the time intervals
}
\description{
The intervals are in sample values of the electrophysiological data.
The direction is based on the direction of movement in 1D environment.
}
