% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{getIntervalsAtHeadDirection}
\alias{getIntervalsAtHeadDirection}
\alias{getIntervalsAtHeadDirection,ANY,ANY-method}
\alias{getIntervalsAtHeadDirection,Positrack-method}
\title{Get time intervals at which the animal's head direction is in a given range}
\usage{
getIntervalsAtHeadDirection(pt, hdMin, hdMax)

\S4method{getIntervalsAtHeadDirection}{Positrack}(pt, hdMin, hdMax)
}
\arguments{
\item{pt}{Positrack object}

\item{hdMin}{Minimal head direction that will be considered (in degree)}

\item{hdMax}{Maximal head direction that will be considered (in degree)}
}
\value{
matrix with the time intervals
}
\description{
Works as though head direction was a linear variable.
}

