% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{getIntervalsAtSpeed}
\alias{getIntervalsAtSpeed}
\alias{getIntervalsAtSpeed,Positrack-method}
\alias{getIntervalsAtSpeed,ANY,ANY-method}
\title{Get time intervals at which the speed of the animal was withing a specified ranged}
\usage{
getIntervalsAtSpeed(pt, minSpeed, maxSpeed)

\S4method{getIntervalsAtSpeed}{Positrack}(pt, minSpeed, maxSpeed)
}
\arguments{
\item{pt}{Positrack object}

\item{minSpeed}{Minimal speed of the animal.}

\item{maxSpeed}{Maximal speed of the animal.}
}
\value{
matrix with the time intervals
}
\description{
The intervals are in sample values of the electrophysiological data
}
