% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{getIntervalsEnvironment}
\alias{getIntervalsEnvironment}
\alias{getIntervalsEnvironment,RecSession-method}
\alias{getIntervalsEnvironment,ANY,ANY-method}
\title{Get the time intervals in sample values for trials in a given environment}
\usage{
getIntervalsEnvironment(rs, environment = "lt")

\S4method{getIntervalsEnvironment}{RecSession}(rs, environment = "lt")
}
\arguments{
\item{rs}{A RecSession object}

\item{environment}{The name of an environment}
}
\value{
matrix with 2 columns containing the start and end of each trial in the environment
}
\description{
Get the time intervals in sample values for trials in a given environment
}
