% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{getIntervalsEnvironment}
\alias{getIntervalsEnvironment}
\alias{getIntervalsEnvironment,ANY,ANY-method}
\alias{getIntervalsEnvironment,RecSession-method}
\title{Get the time intervals in sample values for trials for a given stimulation}
\usage{
getIntervalsEnvironment(rs, stimulation = "lt")

\S4method{getIntervalsEnvironment}{RecSession}(rs, stimulation = "lt")
}
\arguments{
\item{rs}{A RecSession object}

\item{stimulation}{The name of an stimulation}
}
\value{
matrix with 2 columns containing the start and end of each trial for the stimulation
}
\description{
Get the time intervals in sample values for trials in a given stimulation
}

