% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{getIsClustered}
\alias{getIsClustered}
\alias{getIsClustered,ANY,ANY-method}
\alias{getIsClustered,RecSession-method}
\title{Is the recording session clustered?}
\usage{
getIsClustered(rs)

\S4method{getIsClustered}{RecSession}(rs)
}
\arguments{
\item{rs}{A RecSession object}
}
\value{
TRUE or FALSE
}
\description{
Is the recording session clustered?
}

