% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{getIsEarlyProcessed}
\alias{getIsEarlyProcessed}
\alias{getIsEarlyProcessed,RecSession-method}
\alias{getIsEarlyProcessed,ANY,ANY-method}
\title{Has the spike extraction been run on the recording session?}
\usage{
getIsEarlyProcessed(rs)

\S4method{getIsEarlyProcessed}{RecSession}(rs)
}
\arguments{
\item{rs}{A RecSession object}
}
\value{
TRUE or FALSE
}
\description{
Has the spike extraction been run on the recording session?
}
