% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{getMapStats}
\alias{getMapStats}
\alias{getMapStats,ANY,ANY-method}
\alias{getMapStats,SpatialProperties2d-method}
\title{Calculate spatial statistics of the firing rate maps of neurons}
\usage{
getMapStats(sp, st, pt, ...)

\S4method{getMapStats}{SpatialProperties2d}(sp, st, pt,
  border = "rectangular", triggered = FALSE, newMaps = TRUE)
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack object}

\item{border}{Set how the border of the environment will be detected. Value can be set to "rectangular" or "circular. 
Default value is "rectangular"}

\item{triggered}{Logical indicating whether to calculate a spike-triggered firing rate map instead of conventional map
Default value is FALSE}

\item{newMaps}{logical indicating if new maps are calculated when calling the function}
}
\value{
SpatialProperties2d object with the stats in the following slots: peakRate, infoScore, sparsity, borderScore and gridScore
}
\description{
Calculate spatial statistics of the firing rate maps of neurons
}

