% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{getMapStatsShuffle}
\alias{getMapStatsShuffle}
\alias{getMapStatsShuffle,ANY,ANY-method}
\alias{getMapStatsShuffle,SpatialProperties2d-method}
\title{Calculate random spatial statistics of the firing rate maps of neurons}
\usage{
getMapStatsShuffle(sp, st, pt, border = "rectangular", triggered = F)

\S4method{getMapStatsShuffle}{SpatialProperties2d}(sp, st, pt,
  border = "rectangular", triggered = FALSE)
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack object}

\item{border}{Set how the border of the environment will be detected. Value can be set to "rectangular" or "circular.
Default value is rectangular}

\item{triggered}{Logical indicating whether to calculate a spike-triggered firing rate map instead of conventional map
Default value is FALSE}
}
\value{
SpatialProperties2d object with the random stats in the following slots: peakRateShuffle, infoScoreShuffle, 
sparsityShuffle, borderScoreShuffle and gridScoreShuffle, etc.
}
\description{
The random values are obtained by shifting the position data.
The number of shufflings is nShufflings of the SpatialProperties2d object.
}

