% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{methods}
\name{getRecSession}
\alias{getRecSession}
\alias{getRecSession,ElectroProject-method}
\alias{getRecSession,ANY,ANY-method}
\title{Return a RecSession object from the list in an ElectroProject objects}
\usage{
getRecSession(ep, name)

\S4method{getRecSession}{ElectroProject}(ep, name = "")
}
\arguments{
\item{ep}{ElectroProject object}

\item{name}{Name of the RecSession, e.g. jp19841-01072015-0108}
}
\value{
RecSession
}
\description{
Return a RecSession object from the list in an ElectroProject objects
}
