% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{getRecSessionObjects}
\alias{getRecSessionObjects}
\alias{getRecSessionObjects,ANY,ANY-method}
\alias{getRecSessionObjects,RecSession-method}
\title{Load a set of objects that are session specific}
\usage{
getRecSessionObjects(rs)

\S4method{getRecSessionObjects}{RecSession}(rs)
}
\arguments{
\item{rs}{A RecSession object}
}
\value{
a list of objects containing spike trains, position data, data files, cell groups, 
spatial properties, 1D spatial properties, head direction data.
}
\description{
This is used to get the objects that are most commonly needed when doing analysis. 
Instead of creating the object each at a time and having to set some values manually,
just call this function and a list of objects are returned.
}
\details{
If you just want to use one or two objects, the code might run faster if you load what you need manually.
}

