% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{methods}
\name{getSessionList}
\alias{getSessionList}
\alias{getSessionList,ANY,ANY-method}
\alias{getSessionList,ElectroProject-method}
\title{Return a list of RecSession objects that have some common properties}
\usage{
getSessionList(ep, clustered = "", region = "", env = "")

\S4method{getSessionList}{ElectroProject}(ep, clustered = "", region = "",
  env = "")
}
\arguments{
\item{ep}{ElectroProject object}

\item{clustered}{logical indicating whether the session should be clustered or not}

\item{region}{Set to a given brain region to select only sessions with tetrodes in this brain region}

\item{env}{Set to a given environment code to select only sessions during which this environment was presented}
}
\value{
list of RecSession objects
}
\description{
Return a list of RecSession objects that have some common properties
}

