% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{methods}
\name{getSessionListFromSessionNames}
\alias{getSessionListFromSessionNames}
\alias{getSessionListFromSessionNames,ElectroProject-method}
\alias{getSessionListFromSessionNames,ANY,ANY-method}
\title{Return a list of RecSession objects from a character vectors containing session names}
\usage{
getSessionListFromSessionNames(ep, sessionNames)

\S4method{getSessionListFromSessionNames}{ElectroProject}(ep, sessionNames)
}
\arguments{
\item{ep}{ElectroProject object}

\item{sessionNames}{Character vector containing session names}
}
\value{
list of RecSession objects
}
\description{
Return a list of RecSession objects from a character vectors containing session names
}
