% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{getSpeedAtResValues}
\alias{getSpeedAtResValues}
\alias{getSpeedAtResValues,Positrack-method}
\alias{getSpeedAtResValues,ANY,ANY-method}
\title{Get the speed of the animal at given time points}
\usage{
getSpeedAtResValues(pt, res, angular = FALSE)

\S4method{getSpeedAtResValues}{Positrack}(pt, res, angular = FALSE)
}
\arguments{
\item{pt}{Positrack object}

\item{res}{Time values in electrophysiological samples.}

\item{angular}{Logical, if true will use the angular velocity instead of the running speed}
}
\value{
Vector with the speed at the different time points
}
\description{
The time points are in sample values of the electrophysiological data
}
