% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{gridScore}
\alias{gridScore}
\alias{gridScore,SpatialProperties2d-method}
\alias{gridScore,ANY,ANY-method}
\title{Calculate the grid scores from the spatial autocorrelations}
\usage{
gridScore(sp)

\S4method{gridScore}{SpatialProperties2d}(sp)
}
\arguments{
\item{sp}{SpatialProperties2d object}
}
\value{
SpatialProperties2d object with the grid scores in the slot gridScore
}
\description{
Calculate the grid scores from the spatial autocorrelations
}
