% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeadDirection.R
\docType{methods}
\name{hdHistoAsDataFrame}
\alias{hdHistoAsDataFrame}
\alias{hdHistoAsDataFrame,ANY,ANY-method}
\alias{hdHistoAsDataFrame,HeadDirection-method}
\title{Return the hd histos as a data.frame}
\usage{
hdHistoAsDataFrame(hd)

\S4method{hdHistoAsDataFrame}{HeadDirection}(hd)
}
\arguments{
\item{hd}{HeadDirection object}
}
\value{
data.frame with the rate-head direction histograms. Column names are clu.id, deg, rate
}
\description{
Return the hd histos as a data.frame
}

