% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeadDirection.R
\docType{class}
\name{HeadDirection-class}
\alias{HeadDirection-class}
\alias{HeadDirection}
\title{An S4 class used to get head direction histograms and circular statistics for each neuron.}
\description{
The user creates positrack and spike train objects first.
Then the user calls the different methods of this class with the spike train and positrack objects as arguments.
It can get the vector length of the head direction histograms and do shuffling to know the threshold for significance.
}
\section{Slots}{

\describe{
\item{\code{session}}{A charactor object with the name of the recording session}

\item{\code{degPerBin}}{Degrees per bin in the head direction histograms}

\item{\code{smoothOccupancySd}}{The standard deviation of the gaussian kernel used to smooth the occupancy histogram}

\item{\code{smoothRateHistoSd}}{The standard deviation of the gaussian kernel used to smooth the rate histograms}

\item{\code{nBinHisto}}{Number of bins in the histograms}

\item{\code{hdSpikes}}{Head direction of all valid spikes}

\item{\code{histo}}{An array holding the histograms of all neurons}

\item{\code{histoDegree}}{A numeric holding the degree for each bin of the histogram}

\item{\code{occupancy}}{A numeric holding the occupancy histogram}

\item{\code{cellList}}{A list of cluster id of the neurons}

\item{\code{cellPairList}}{Data frame containing pairs of cells}

\item{\code{histoRepetitions}}{Number of repetition of the 0-360 range in the histogram. Default is 0 repetition}

\item{\code{peakRates}}{Peak firing rate in Hz in each firing rate histogram}

\item{\code{vectorLength}}{Mean vector length of each firing rate histogram}

\item{\code{meanDirection}}{Mean direction in each firing rate histogram}

\item{\code{nShufflings}}{Number of shufflings to get a distribution of vector length that would be obtained by chance}

\item{\code{minShiftMs}}{Minimum time shift of the head direction data used during the shuffling procedure}

\item{\code{peakRatesShuffle}}{Numeric to hold the peak firing rates obtained during the shuffling procedure}

\item{\code{vectorLengthShuffle}}{Numeric to hold the vector length obtained during the shuffling procedure}

\item{\code{crossHisto}}{Array containing crosscorrelation rate histogram}
}}

