% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeadDirection.R
\docType{methods}
\name{headDirectionHisto}
\alias{headDirectionHisto}
\alias{headDirectionHisto,HeadDirection-method}
\alias{headDirectionHisto,ANY,ANY-method}
\title{Calculate the head direction rate histograms of neurons using a HeadDirection, SpikeTrain and Positrack objects}
\usage{
headDirectionHisto(hd, st, pt)

\S4method{headDirectionHisto}{HeadDirection}(hd, st, pt)
}
\arguments{
\item{hd}{HeadDirection object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack}
}
\value{
HeadDirection object with the firing rate histograms in the slot histo
}
\description{
Calculate the head direction rate histograms of neurons using a HeadDirection, SpikeTrain and Positrack objects
}
