% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsSpatial.R
\name{headDirectionPolarPlot}
\alias{headDirectionPolarPlot}
\title{Polar plot of firing rate as a function of head direction}
\usage{
headDirectionPolarPlot(histo, outma = c(0.5, 0.5, 0.5, 0.5), margin = c(0.5,
  0.3, 0.5, 0.3), axis.x.mgp = c(1, 0.3, 0), axis.y.mgp = c(2.2, 0.6, 0),
  cex.x.axis = 0.5, cex.y.axis = 0.5, cex.lab = 0.5, xlab = "",
  ylab = "", show.xlab = TRUE, main.title = "", peak.rate.prefix = "",
  clockwise = TRUE)
}
\arguments{
\item{histo}{A data.frame with columns deg and rate, or a numeric vector containing rate values}

\item{outma}{Outer margins of the figure}

\item{margin}{Inner margins of the figure}

\item{axis.x.mgp}{mgp for x axis}

\item{axis.y.mgp}{mgp for y axis}

\item{cex.x.axis}{cex for x axis}

\item{cex.y.axis}{cex for y axis}

\item{cex.lab}{cex for labels}

\item{xlab}{Name to display under the x axis}

\item{ylab}{Name to display at the left of the y axis}

\item{show.xlab}{Wether of not the xlabel is shown}

\item{main.title}{A title for the figure}

\item{peak.rate.prefix}{Additional information to display before the peak firing rate.}

\item{clockwise}{plot degrees clockwise or anticlockwise}
}
\description{
Polar representation of firing rate using the polar.plot function of the plotrix package
}
