% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsSpatial.R
\name{headDirectionPolarPlots}
\alias{headDirectionPolarPlots}
\title{Plot several polar plot with firing rate as a function of head direction}
\usage{
headDirectionPolarPlots(histo, names, fn = "page.full.plot.pdf")
}
\arguments{
\item{histo}{A matrix [deg,clu], or numeric if there is only one cell, containing the data for the polar plots.}

\item{names}{A character vector containing the name of each plot}

\item{fn}{Character vector containing the file name for the plot}
}
\description{
Plot several polar plot with firing rate as a function of head direction
}

