% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeadDirection.R
\docType{methods}
\name{headDirectionStatsShuffle}
\alias{headDirectionStatsShuffle}
\alias{headDirectionStatsShuffle,HeadDirection-method}
\alias{headDirectionStatsShuffle,ANY,ANY-method}
\title{Calculate the random head direction statistics for the head direction rate histograms with 
a HeadDirection, SpikeTrain and Positrack objects}
\usage{
headDirectionStatsShuffle(hd, st, pt)

\S4method{headDirectionStatsShuffle}{HeadDirection}(hd, st, pt)
}
\arguments{
\item{hd}{HeadDirection object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack}
}
\value{
HeadDirection object with the random statistics in the slots 
peakRatesShuffle, vectorLengthShuffle and meanDirectionShuffle
}
\description{
Calculate the random head direction statistics for the head direction rate histograms with 
a HeadDirection, SpikeTrain and Positrack objects
}
