% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties1d.R
\docType{methods}
\name{histoStatsAsDataFrame}
\alias{histoStatsAsDataFrame}
\alias{histoStatsAsDataFrame,SpatialProperties1d-method}
\alias{histoStatsAsDataFrame,ANY,ANY-method}
\title{Return the histo stats as data.frame.}
\usage{
histoStatsAsDataFrame(sp1, shuffle)

\S4method{histoStatsAsDataFrame}{SpatialProperties1d}(sp1, shuffle = FALSE)
}
\arguments{
\item{sp1}{SpatialProperties1d object}

\item{shuffle}{Logical, if TRUE will return the statistics from the shuffle.}
}
\value{
data.frame with the statistics of the firing rate histograms
}
\description{
The column names in the data.frame are clu.id, x, rate.
}
