% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{identifySpikeTimes}
\alias{identifySpikeTimes}
\title{Detect spike time using filtered signal and root mean square arrays}
\usage{
identifySpikeTimes(dataf, power, powerWindowSize, powerWindowSlide,
  powerThreshold)
}
\arguments{
\item{dataf}{Filtered brain signal}

\item{power}{Power in sliding time windows}

\item{powerWindowSize}{Size of the power window}

\item{powerWindowSlide}{Amount of shift in the power window between subsequent power estimates}

\item{powerThreshold}{Threshold used to detect spikes}
}
\value{
Matrix containing spike times and negative peak value of the spike. The spike times are in sample index starting at index 0.
}
\description{
Detect spike time using filtered signal and root mean square arrays
}

