% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{ifr}
\alias{ifr}
\alias{ifr,ANY,ANY-method}
\alias{ifr,SpikeTrain-method}
\title{Calculate the instantaneous firing rate from the spike trains.}
\usage{
ifr(st, windowSizeMs = 100, spikeBinMs = 1, kernelSdMs = 100)

\S4method{ifr}{SpikeTrain}(st, windowSizeMs = 100, spikeBinMs = 1,
  kernelSdMs = 100)
}
\arguments{
\item{st}{SpikeTrain object}

\item{windowSizeMs}{The window size for each ifr value}

\item{spikeBinMs}{The bin size for the spike count array}

\item{kernelSdMs}{Standard deviation of the gaussian kernel used to smooth the spike count vector}
}
\value{
SpikeTrain object with the instantaneous firing rate
}
\description{
The ifr for the periods within the intervals set within the SpikeTrain object will be calculated.
A vector with the spike count in each bin is calculated
Then a gaussian kernel is applied to the spike count vector
Finally, the firing probability is integrated over a set window size and transform to a firing rate.
}

