% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{ifrAssociation}
\alias{ifrAssociation}
\alias{ifrAssociation,SpikeTrain-method}
\alias{ifrAssociation,ANY,ANY-method}
\title{Calculate the correlation coefficient between the instantaneous firing rate of the neurons.}
\usage{
ifrAssociation(st)

\S4method{ifrAssociation}{SpikeTrain}(st)
}
\arguments{
\item{st}{SpikeTrain object with ifr already calculated}
}
\value{
data.frame with the clu.ids of pairs and correlation coefficient between the instantaneous firing rate.
}
\description{
Call ifr(st) method before calling this function.
Only the ifr falling withing the intervals set in the SpikeTime object will be considered.
}
