% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{isolationDistance}
\alias{isolationDistance}
\alias{isolationDistance,SpikeTrain-method}
\alias{isolationDistance,ANY,ANY-method}
\title{Get the isolation distance of each cluster}
\usage{
isolationDistance(st, cg)

\S4method{isolationDistance}{SpikeTrain}(st, cg)
}
\arguments{
\item{st}{SpikeTrain object}

\item{cg}{CellGroup object}
}
\value{
a SpikeTrain object with the isolationDistance set.
}
\description{
For each cluster, the mahalanobis distance of all spikes relative to the cluster
is calculated. The isolation distance is the minimal distance from the cluster center
at which there are as many spikes from other cluster than from the cluster of interest.
Distance is not defined for cases in which the number of cluster spikes is
greater than the number of noise spikes.
}
\details{
From Schmitzer-Torbert et al. 2005
}
