% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{linearzeLinearTrack}
\alias{linearizeLinearTrack,ANY,ANY-method}
\alias{linearzeLinearTrack}
\alias{linearzeLinearTrack,Positrack-method}
\title{Linearize the position data}
\usage{
linearzeLinearTrack(pt)

\S4method{linearzeLinearTrack}{Positrack}(pt)
}
\arguments{
\item{pt}{Positrack object}
}
\value{
Positrack object with linear position (lin)
}
\description{
This is usually done if the animal ran on a linear track. 
The regression line of the position data is calculated and the 
closest point on the line is found for each 2D coordinate.
}

