% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsSpatial.R
\name{linearRatePlot}
\alias{linearRatePlot}
\title{Plot a linear rate histogram from a SpatialProperties1D object}
\usage{
linearRatePlot(sp1, n = 1, outma = c(1, 1, 1, 0), margin = c(1.5, 2, 0.8,
  0.3), axis.x.pos = 0, axis.y.pos = 0, axis.y.las = 2, mgp.x = c(0.5,
  0.05, 0), mgp.y = c(0.8, 0.3, 0.2), xlab = "Position (cm)",
  ylab = "Rate (Hz)", xaxis.at = seq(0, 80, 20), yaxis.at = seq(0, 50,
  10), ...)
}
\arguments{
\item{sp1}{SpatialProperties1d object}

\item{n}{Index of the histogram in the sp1 object that you want to plot}

\item{outma}{Outer margins of the figure}

\item{margin}{Inner margins of the figure}

\item{axis.x.pos}{position of x axis}

\item{axis.y.pos}{position of y axis}

\item{axis.y.las}{las of y axis}

\item{mgp.x}{mgp for x axis}

\item{mgp.y}{mgp for y axis}

\item{xlab}{Name to display under the x axis}

\item{ylab}{Name to display at the left of the y axis}

\item{xaxis.at}{where to put the tics}

\item{yaxis.at}{where to put the tics}

\item{...}{passed to the plot function}
}
\description{
In development
}

