% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsSpatial.R
\name{linearRatePlots}
\alias{linearRatePlots}
\title{Plot several linear rate histograms from a SpatialProperties1D object}
\usage{
linearRatePlots(sp1, ncol = 5, nrow = 6)
}
\arguments{
\item{sp1}{SpatialProperties1d object}

\item{ncol}{Number of columns of plots per page}

\item{nrow}{Number of rows of plots per page}
}
\description{
In development
}
