% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellGroup.R
\docType{methods}
\name{loadCellGroup}
\alias{loadCellGroup}
\alias{loadCellGroup,CellGroup-method}
\alias{loadCellGroup,ANY,ANY-method}
\title{Load the information regarding a group of cells}
\usage{
loadCellGroup(cg)

\S4method{loadCellGroup}{CellGroup}(cg)
}
\arguments{
\item{cg}{A CellGroup object}
}
\value{
A CellGroup object with the information loaded
}
\description{
The object will get the tetrode id and
cluster number on the tetrode for each cluster.
The object also load the brain region associated with each cell.
This infomration is read from the .desel file and the clu file
of each tetrode.
}
