% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{loadPositrack}
\alias{loadPositrack}
\alias{loadPositrack,ANY,ANY-method}
\alias{loadPositrack,Positrack-method}
\title{Load position data from the .whl and .whd files}
\usage{
loadPositrack(pt)

\S4method{loadPositrack}{Positrack}(pt)
}
\arguments{
\item{pt}{Positrack object}
}
\value{
Positrack object
}
\description{
This function also use the .res_samples_per_whl_sample, .sampling_rate_dat and .px_per_cm files
Some smoothing is done on the x and y data.
Slots x and y are in cm.
}

