% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{methods}
\name{loadSessionsInList}
\alias{loadSessionsInList}
\alias{loadSessionsInList,ElectroProject-method}
\alias{loadSessionsInList,ANY,ANY-method}
\title{Create a List of RecSessions objects from the SessionNameList}
\usage{
loadSessionsInList(ep)

\S4method{loadSessionsInList}{ElectroProject}(ep)
}
\arguments{
\item{ep}{ElectroProject object}
}
\value{
ElectroProject object with RecSession loaded
}
\description{
The RecSession objects will be initialized.
}
