% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{loadSpikeTrain}
\alias{loadSpikeTrain}
\alias{loadSpikeTrain,ANY,ANY-method}
\alias{loadSpikeTrain,SpikeTrain-method}
\title{Load the spike train from the .clu and .res files}
\usage{
loadSpikeTrain(st)

\S4method{loadSpikeTrain}{SpikeTrain}(st)
}
\arguments{
\item{st}{SpikeTrain object}
}
\value{
SpikeTrain object with the instantaneous firing rate
}
\description{
Load the spike train from the .clu and .res files
}

