% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{mapsAsDataFrame}
\alias{mapsAsDataFrame}
\alias{mapsAsDataFrame,ANY,ANY-method}
\alias{mapsAsDataFrame,SpatialProperties2d-method}
\title{Return the firing rate maps in a data.frame}
\usage{
mapsAsDataFrame(sp)

\S4method{mapsAsDataFrame}{SpatialProperties2d}(sp)
}
\arguments{
\item{sp}{SpatialProperties2d object}
}
\value{
data.frame with the firing rate maps. Column names are clu.id, x, y, rate
}
\description{
Return the firing rate maps in a data.frame
}

