% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{mapSpatialAutocorrelation}
\alias{mapSpatialAutocorrelation}
\alias{mapSpatialAutocorrelation,SpatialProperties2d-method}
\alias{mapSpatialAutocorrelation,ANY,ANY-method}
\title{Calculate the spatial autocorrelation of the firing rate maps of neurons}
\usage{
mapSpatialAutocorrelation(sp)

\S4method{mapSpatialAutocorrelation}{SpatialProperties2d}(sp)
}
\arguments{
\item{sp}{SpatialProperties2d object}
}
\value{
SpatialProperties2d object with the spatial autocorrelation in slot auto
}
\description{
The autocorrelation is performed on the firing rate maps of the SpatialProperties2d object
}
