% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{meanFiringRate}
\alias{meanFiringRate}
\alias{meanFiringRate,ANY,ANY-method}
\alias{meanFiringRate,SpikeTrain-method}
\title{Calculate the mean firing rate (Hz) of each neuron in a SpikeTrain object}
\usage{
meanFiringRate(st)

\S4method{meanFiringRate}{SpikeTrain}(st)
}
\arguments{
\item{st}{SpikeTrain object}
}
\value{
a SpikeTrain object with the mean firing rate in slot meanFiringRate
}
\description{
This is simply the number of spikes divided by the time within the intervals
set in the SpikeTrain object.
}

