% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{mergeSimultaneousSpikes}
\alias{mergeSimultaneousSpikes}
\title{Merge simulatenous spikes that were detected on different wires}
\usage{
mergeSimultaneousSpikes(time, trough, maxTimeDifference)
}
\arguments{
\item{time}{Numeric containing the spike times}

\item{trough}{Numeric containing the trough of detected spikes}

\item{maxTimeDifference}{Maximal time difference to be considered simultaneous}
}
\value{
Numeric containing the spike times
}
\description{
The spike time associated with the smallest trough is kept.
}
