% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{mergeTetrodeSpecificResCluFiles}
\alias{mergeTetrodeSpecificResCluFiles}
\title{Merge tetrode specific clu and res file into a main res and clu file}
\usage{
mergeTetrodeSpecificResCluFiles(rs)
}
\arguments{
\item{rs}{RecSession}
}
\description{
This is run after the clustering of all tetrodes is done. It create
the main .res and .clu files that are used when doing subsequent 
analysis. Cluster 1 is noise. 
The clusters 1 from different tetrodes are merged. The other
clusters are not merged. They will be numbered from 2 to N, where N is 
the total number of clusters in the recording session.
}

