% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{Mode}
\alias{Mode}
\title{Get the mode of a distribution}
\usage{
Mode(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
The value in the vector with the highest frequency
}
\description{
This get you the value with the highest frequency in a vector
}

