% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{percentileFiringRateMaps}
\alias{percentileFiringRateMaps}
\alias{percentileFiringRateMaps,ANY,ANY-method}
\alias{percentileFiringRateMaps,SpatialProperties2d-method}
\title{Get a given percentile from the firing rate maps of a SpatialProperties2d object}
\usage{
percentileFiringRateMaps(sp, percentile)

\S4method{percentileFiringRateMaps}{SpatialProperties2d}(sp, percentile = 75)
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{percentile}{Given percentile to be retrieved}
}
\value{
Numeric vector with the xth percentile of the firing rate maps
}
\description{
Get a given percentile from the firing rate maps of a SpatialProperties2d object
}

