% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsGeneric.R
\name{plotPoints}
\alias{plotPoints}
\title{Plot points in 2d}
\usage{
plotPoints(data, v1 = "v1", v2 = "v2", axis.y.pos = -0.2,
  axis.x.pos = -0.2, axis.y.las = 2, main.title = "", mgp.x = c(0.5,
  0.1, 0.1), mgp.y = c(0.9, 0.2, 0.1), xlab = "", ylab = "",
  plotxlim = c(-0.2, 0.6), plotylim = c(-0.2, 0.6), outma = c(0.5, 0.5,
  0.5, 0.5), margin = c(1.5, 1.5, 1, 0.3), xaxis.at = seq(-0.2, 0.7, 0.2),
  yaxis.at = seq(-0.2, 0.7, 0.2), cex.point = 0.1, add.text = "",
  add.text.pos = c(0, 0.5), ...)
}
\arguments{
\item{data}{data.frame with the data}

\item{v1}{Name of the variable to plot on the x axis}

\item{v2}{Name of the variable to plot on the y axis}

\item{axis.y.pos}{Position of the y axis}

\item{axis.x.pos}{Position of the x axis}

\item{axis.y.las}{Orientation of the letters for the label of the y axis}

\item{main.title}{Main title}

\item{mgp.x}{mgp for x axis}

\item{mgp.y}{mgp for y axis}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{plotxlim}{Limit of the x axis}

\item{plotylim}{Limit of the y axis}

\item{outma}{outma for setting par}

\item{margin}{margin for setting par}

\item{xaxis.at}{Where the tics are shown on the x axis}

\item{yaxis.at}{Where the tics are shown on the y axis}

\item{cex.point}{cex for the size of the points}

\item{add.text}{Text to add to the plot}

\item{add.text.pos}{Position of the text to add, format (x,y)}

\item{...}{Passed to the graphics::plot function}
}
\description{
Just a wrapper around plot and points functions.
}

