% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{plotSimulatedRawTrace}
\alias{plotSimulatedRawTrace}
\title{Plot the raw traces simulation generated by the simulateRawTrace function}
\usage{
plotSimulatedRawTrace(sim, detectedSpikeTimes = NA, ...)
}
\arguments{
\item{sim}{list containing the trace, cluId and spikeTime}

\item{detectedSpikeTimes}{if you want to plot detected spikes}

\item{...}{passed to the plot function}
}
\description{
Plot the raw traces simulation generated by the simulateRawTrace function
}
