% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{plotSpikeDetectionSegment}
\alias{plotSpikeDetectionSegment}
\title{Plot the raw traces of a tetrode together with spike times of real data}
\usage{
plotSpikeDetectionSegment(rs, df, tetrodeNo = 1, minPassHz = 800,
  maxPassHz = 8000, firstSample = 0, lastSample = 5000, addSpikes = NA,
  powerWindowSizeMs = 0.3, powerWindowSlideMs = 0.1, SDThreshold = 2.25,
  spikeDetectionRefractoryMs = 0.5)
}
\arguments{
\item{rs}{RecSession object}

\item{df}{DatFiles object}

\item{tetrodeNo}{Numeric indicating the Tetrode number}

\item{minPassHz}{Min frequency for the band pass filter}

\item{maxPassHz}{Max frequency for the band pass filter}

\item{firstSample}{First sample to display}

\item{lastSample}{Last sample to display}

\item{addSpikes}{Numeric containing spikes as a comparison for spikes in .res.tetrodeNo file.}

\item{powerWindowSizeMs}{Power window}

\item{powerWindowSlideMs}{Slide of the power window.}

\item{SDThreshold}{Power threshold}

\item{spikeDetectionRefractoryMs}{Refractory in the spike detection}
}
\description{
This function can be used to visualise the detected spikes on the raw traces.
Use this to get a qualitative impression of the spike detection.
}
\details{
Filtering might create some artefacts at the very beginning and end of traces.
}
