% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{plotSpikes}
\alias{plotSpikes}
\title{Plot the waveform of spikes
 
 Use the fil file for each channel}
\usage{
plotSpikes(rs, channels, res, windowMs = 2)
}
\arguments{
\item{rs}{RecSession}

\item{channels}{Numeric with the list of channels for the tetrode on which the spikes were detected}

\item{res}{Numeric with the time stamps of the spikes in sample value}

\item{windowMs}{Number of ms that will be shown}
}
\description{
Plot the waveform of spikes
 
 Use the fil file for each channel
}

