% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{class}
\name{Positrack-class}
\alias{Positrack-class}
\alias{Positrack}
\title{An S4 class representing the path of an animal during a recording session}
\description{
This class is used to manipulate and represent the position data of the animal.
The first data point is at time resSamplesPerWhlSample and not 0.
The data are usually loaded from .whd, .res_samples_per_whl_sample, .sampling_rate_dat and .px_per_cm
}
\section{Slots}{

\describe{
\item{\code{session}}{A character vector containing the names of the recording session.}

\item{\code{path}}{The directory in which the files of the session are located.}

\item{\code{pxPerCm}}{Pixels per centimeter in the x and y position data.}

\item{\code{samplingRateDat}}{Sampling rate of the .dat files in this recording session}

\item{\code{resSamplesPerWhlSample}}{Number of samples in the .dat files between each position values.}

\item{\code{xWhl}}{A numeric vector containing x data loaded from the .whl file.}

\item{\code{yWhl}}{A numeric vector containing y data loaded from the .whl file.}

\item{\code{hdWhd}}{A numeric vector containing the head direction of the animal, loaded from .whd file.}

\item{\code{x}}{x position of the animal in cm}

\item{\code{y}}{y position of the animal in cm}

\item{\code{hd}}{Head direction of the animal}

\item{\code{lin}}{Linearized position of the animal. Used for linear track data.}

\item{\code{dir}}{Direction in the linearized position data (0 or 1)}

\item{\code{speed}}{Linear speed in cm per sec}

\item{\code{angularSpeed}}{Angular speed in degrees per sec}

\item{\code{res}}{Sample number associated with each position value}

\item{\code{defaultXYSmoothing}}{Default smoothing apply to the x and y position.}

\item{\code{minShiftMs}}{Minimum shift of the position vector during a shuffling procedure}
}}

