% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalProcessing.R
\name{powerRootMeanSquare}
\alias{powerRootMeanSquare}
\title{Calculate the root mean square within a sliding window}
\usage{
powerRootMeanSquare(data, windowSizeSamples, windowSlide)
}
\arguments{
\item{data}{Numeric vector containing the signal}

\item{windowSizeSamples}{Window size in number of samples within which the power will be calculated}

\item{windowSlide}{Shift of the window between calculation of root mean square}
}
\value{
Numeric, root mean square of all windows
}
\description{
Calculate the root mean square within a sliding window
}

