% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties1d.R
\docType{methods}
\name{rateHistoAsDataFrame}
\alias{rateHistoAsDataFrame}
\alias{rateHistoAsDataFrame,ANY,ANY-method}
\alias{rateHistoAsDataFrame,SpatialProperties1d-method}
\title{Return the firing rate histograms as a data.frame.}
\usage{
rateHistoAsDataFrame(sp1)

\S4method{rateHistoAsDataFrame}{SpatialProperties1d}(sp1)
}
\arguments{
\item{sp1}{SpatialProperties1d object}
}
\value{
data.frame with the firing rate histograms
}
\description{
The column names in the data.frame are clu.id, x, rate.
}

