% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{readFetFile}
\alias{readFetFile}
\title{Read a fet file}
\usage{
readFetFile(fileName)
}
\arguments{
\item{fileName}{File name}
}
\value{
A matrix with the spike features, one spike per row
}
\description{
Read a file with the spike features. Uses a c function for speed.
}

