% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{recordingDate}
\alias{recordingDate}
\alias{recordingDate,ANY,ANY-method}
\alias{recordingDate,RecSession-method}
\title{Get the recording date of a recSession, taken from session name}
\usage{
recordingDate(rs)

\S4method{recordingDate}{RecSession}(rs)
}
\arguments{
\item{rs}{A RecSession object}
}
\value{
Date
}
\description{
Get the recording date of a recSession, taken from session name
}

