% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{class}
\name{RecSession-class}
\alias{RecSession}
\alias{RecSession-class}
\title{A S4 class to represent a recording session.}
\description{
A RecSession object contains a description of the recording session.
}
\details{
Sessions are usually divided in several trials with one .dat file for each trial.
}
\section{Slots}{

\describe{
\item{\code{session}}{Name of the recording session}

\item{\code{path}}{Directory where the recording session data are located}

\item{\code{fileBase}}{Filebase of the session}

\item{\code{animalName}}{Name of the animal}

\item{\code{samplingRate}}{Sampling rate of the electrophysiological data}

\item{\code{resofs}}{Number of samples in each trial}

\item{\code{env}}{List of environment for each trial}

\item{\code{electrodeLocation}}{List of electrode location, one per electrode}

\item{\code{trialStartRes}}{Sample at which a trial starts.}

\item{\code{trialEndRes}}{Sample at which a trial ends.}

\item{\code{trialNames}}{Filebase of the individual trials}

\item{\code{trialDurationSec}}{Length of the trials in sec}

\item{\code{sessionDurationSec}}{Total length of the session in sec}

\item{\code{nElectrodes}}{Number of electrodes}

\item{\code{nChannels}}{Number of channels in the electrophysiological recordings}

\item{\code{nTrials}}{Number of trials in the session}

\item{\code{channelsTetrode}}{Matrix containing the channel numbers associated with each tetrode}

\item{\code{clustered}}{Logical indicating if the spikes are clustered}

\item{\code{earlyProcessed}}{Logical indicating if spike extraction has been done}
}}

