% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsGeneric.R
\name{RecSessionListTrialPlot}
\alias{RecSessionListTrialPlot}
\title{Plot function to display the recording trials of each recording session in a sessionList}
\usage{
RecSessionListTrialPlot(sessionList, outma = c(1, 1, 0.5, 0.5),
  margin = c(1.5, 1.5, 1, 0.3), mgp.x = c(0.4, 0.2, 0), mgp.y = c(1, 0.2,
  0))
}
\arguments{
\item{sessionList}{List of RecSession objects}

\item{outma}{outma for setting par}

\item{margin}{margin for setting par}

\item{mgp.x}{mgp.x for setting par of xaxis}

\item{mgp.y}{mgp.y for setting par of yaxis}
}
\description{
Easy way to visualize the environment used in each trial of a list of recording sessions
}
