% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{refractoryRatio}
\alias{refractoryRatio}
\alias{refractoryRatio,ANY,ANY-method}
\alias{refractoryRatio,SpikeTrain-method}
\title{Get refractory ratio of each cluster from its spike-time autocorrelation}
\usage{
refractoryRatio(st, ...)

\S4method{refractoryRatio}{SpikeTrain}(st, refractoryMs = 1.5,
  binSizeMs = 0.5, windowSizeMs = 25, minControlWindowMs = 5,
  maxControlWindowMs = 25)
}
\arguments{
\item{st}{SpikeTrain object}

\item{refractoryMs}{Length of the refractory period in ms}

\item{binSizeMs}{Size of the bins in the spike-time autocorrelation}

\item{windowSizeMs}{Size of the window used to construct the spike-time autocorrelation}

\item{minControlWindowMs}{Minimum time of the control window}

\item{maxControlWindowMs}{Maximal time of the control window}
}
\value{
a SpikeTrain object with the refractoryRation set.
}
\description{
This is the ratio between the mean number of spikes falling in the bins of the refractory period
compared to the max number of spikes falling in one bin of the control period outside the refractory period.
Note that the spike-time autocorrelations in the SpikeTrain object will be modified.
}

