\name{relectro-package}
\alias{relectro-package}
\alias{relectro}
\docType{package}
\title{
What the package does (short line)

}
\description{
More about what it does (maybe more than one line)

}
\details{
\tabular{ll}{
Package: \tab relectro\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-02-28\cr
License: \tab What license is it under?\cr
Depends: \tab methods\cr
}
 An overview of how to use the package, including the most important functions
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>

}
\references{
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{



}
