% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{methods}
\name{runOnSessionList}
\alias{runOnSessionList}
\alias{runOnSessionList,ANY,ANY-method}
\alias{runOnSessionList,ElectroProject-method}
\title{Running a function on a set of recording sessions}
\usage{
runOnSessionList(ep, sessionList, fnct = function(x) {     NA }, save = T,
  overwrite = T, parallel = F, cluster = "", ...)

\S4method{runOnSessionList}{ElectroProject}(ep, sessionList,
  fnct = function(x) {     NA }, save = T, overwrite = T, parallel = F,
  cluster = "", ...)
}
\arguments{
\item{ep}{ElectroProject object}

\item{sessionList}{List of RecSession objects on which the function will be applied}

\item{fnct}{A function to run on each RecSession}

\item{save}{Whether you want to save the data returned by the function}

\item{parallel}{Whether you want to run the function in parallel}

\item{cluster}{A cluster generated from the makeCluster function of the snow package}

\item{...}{optional arguments to fnct}

\item{overwirte}{Logical indicating if the data returned by the function will overwrite the old one when saving into a file}
}
\value{
list of list returned by the lapply function
}
\description{
This applies a function to a list of RecSession objects.
If save is set to TRUE, the results returned by the function will be saved. 
The function applied to single recording sessions should return the results in a list. 
The results are saved in the resultsDirectory of the ElectroProject object.
The names of the files saved will be the name of the elements in the list returned by the function
The data from each recording session will be bound togheter using rbind. If the element is an array, they will be bound with abind.
You can use snow::parLapply instead of lapply by setting parallel to TRUE and passing a valid cluster to the function
If save is set to FALSE, the data returned by the lapply function will not be saved but retured.
If overwrite is set to TRUE, previous data will be overwrite when saving the data.
If overwrite is set to FALSE, the new data will be appended to the old one.
}

