% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{sem}
\alias{sem}
\title{Get the standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric, standard error of the mean
}
\description{
The standard deviation divided by the square root of n
}
