% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\name{sessionNameFromCluId}
\alias{sessionNameFromCluId}
\title{Get session name from cluId}
\usage{
sessionNameFromCluId(cluId = NULL)
}
\arguments{
\item{cluId}{Character vector with the session name}
}
\value{
Character vector with session name
}
\description{
Assumes the session name is in the format name-date-rest and cluId is name-data-rest_cluNo
}
