% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\name{sessionNamesFromSessionList}
\alias{sessionNamesFromSessionList}
\title{Get a character vector with the session names of a list of RecSession object}
\usage{
sessionNamesFromSessionList(rss)
}
\arguments{
\item{rsl}{List containing recording sessions}
}
\value{
Charcter vector of the session names
}
\description{
Get a character vector with the session names of a list of RecSession object
}

