% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{setEvents}
\alias{setEvents}
\alias{setEvents,ANY,ANY-method}
\alias{setEvents,SpikeTrain-method}
\title{Set some events for spike-time crosscorrelation to the events.}
\usage{
setEvents(st, events)

\S4method{setEvents}{SpikeTrain}(st, events = NULL)
}
\arguments{
\item{st}{SpikeTrain object}

\item{events}{Time in sample number of the events}
}
\value{
a SpikeTrain object with the events set
}
\description{
These events could be laser stimulation or some behavioural events
}

