% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{setIntervals}
\alias{setIntervals}
\alias{setIntervals,ANY,ANY-method}
\alias{setIntervals,SpikeTrain-method}
\title{Set time intervals to limit the period used in the analysis}
\usage{
setIntervals(st, s, e)

\S4method{setIntervals}{SpikeTrain}(st, s, e)
}
\arguments{
\item{st}{SpikeTrain object}

\item{s}{Vector or matrix. If a matrix, should have 2 columns (beginning and end of intervals).
If a vector, beginning of intervals.}

\item{e}{Vector, end of the intervals. If s is a matrix, e is not used.}
}
\value{
a SpikeTrain object with the intervals set.
}
\description{
Only the data within the intervals are used for analysis.
For example with interval 0-20000, a spike at time 0 or 20000 is not included.
Spikes between time 0 and 20000 are used.
These intervals are used in SpikeTrain methods and also in
methods of other classes (SaptialProperties2d, SpatialProperties1d, HeadDirection, etc.).
By default, the intervals are set from 0 to time point of the last recorded spike.
}

