% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{setInvalidOutsideInterval}
\alias{setInvalidOutsideInterval}
\alias{setInvalidOutsideInterval,ANY,ANY-method}
\alias{setInvalidOutsideInterval,Positrack-method}
\title{Set position data outside time intervals to NA}
\usage{
setInvalidOutsideInterval(pt, s, e)

\S4method{setInvalidOutsideInterval}{Positrack}(pt, s, e = "")
}
\arguments{
\item{pt}{Positrack object}

\item{s}{Vector or matrix, If matrix, then should have 2 columns, for beginning and end of intervals. 
If vector, then it is the beginning of intervals.}

\item{e}{Vector, end of the intervals. Not used if s is a matrix}
}
\value{
Positrack object with only valid value within the intervals.
}
\description{
The slots x, y, hd, speed, lin and dir are affected.
The intervals are in sample values of the electrophysiological data
}

