% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{setPositrack}
\alias{setPositrack}
\alias{setPositrack,Positrack-method}
\alias{setPositrack,ANY,ANY-method}
\title{set position values from arguments}
\usage{
setPositrack(pt, pxX, pxY, hd, resSamplesPerWhlSample, samplingRateDat, pxPerCm)

\S4method{setPositrack}{Positrack}(pt, pxX, pxY, hd, resSamplesPerWhlSample,
  samplingRateDat, pxPerCm)
}
\arguments{
\item{pt}{Positrack object}

\item{pxX}{Numeric with the x position of animal. x is in pixels. invalid is -1.0}

\item{pxY}{Numeric with the y position of animal. y is in pixels. invalid is -1.0}

\item{hd}{Numeric with the head direction of animal in degrees. invalid is -1.0}

\item{resSamplesPerWhlSample}{Number of electrophysiological samples per position sample}

\item{samplingRateDat}{Sampling rate of electrophysiological data}

\item{pxPerCm}{Number of pixels per cm}
}
\value{
Positrack object
}
\description{
This function also use the .res_samples_per_whl_sample, .sampling_rate_dat and .px_per_cm files
}
