% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{methods}
\name{setSessionList}
\alias{setSessionList}
\alias{setSessionList,ElectroProject-method}
\alias{setSessionList,ANY,ANY-method}
\title{Create a list of RecSession objects for the ElectroProject}
\usage{
setSessionList(ep, loadSessions = TRUE)

\S4method{setSessionList}{ElectroProject}(ep, loadSessions = TRUE)
}
\arguments{
\item{ep}{ElectroProject object}

\item{loadSessions}{logical, whether the RecSession object should be created}
}
\value{
ElectroProject object with a list of session names and directories
}
\description{
Will list directories in the project directories.
Only directories with an hyphen in their names are considered recSession directories.
So only recording session directories should have a hyphen in their names
}
