% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{setSpikeTrain}
\alias{setSpikeTrain}
\alias{setSpikeTrain,SpikeTrain-method}
\alias{setSpikeTrain,ANY,ANY-method}
\title{Set new spike trains}
\usage{
setSpikeTrain(st, res, clu, samplingRate)

\S4method{setSpikeTrain}{SpikeTrain}(st, res, clu, samplingRate)
}
\arguments{
\item{st}{SpikeTrain object}

\item{res}{Spike times in sample number}

\item{clu}{Cluster id for each spike}

\item{samplingRate}{The number of spamples per second (Hz)}
}
\value{
SpikeTrain object with the instantaneous firing rate
}
\description{
This function is used to set spike trains
}
