% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{shift}
\alias{shift}
\title{Shift values in a vector by a certain number of places and in a given direction}
\usage{
shift(v, places, dir = "right")
}
\arguments{
\item{v}{A vector}

\item{places}{Number of places the values will be moved.}

\item{dir}{Direction of the shift, values should be right or left (or r or l).}
}
\description{
The vector is wrapped around so that values that would end up after the end of the vector are place at the beginning.
}
\examples{
shift(v=1:10, place=2, dir="r")
}
