% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{shiftHdRandom}
\alias{shiftHdRandom}
\alias{shiftHdRandom,ANY,ANY-method}
\alias{shiftHdRandom,Positrack-method}
\title{Shift the head direction data of the animal of a random amount larger than the value of minShiftMs}
\usage{
shiftHdRandom(pt)

\S4method{shiftHdRandom}{Positrack}(pt)
}
\arguments{
\item{pt}{Positrack object}
}
\value{
Positrack object with shifted head direction (hd) values
}
\description{
This is used to do shuffling. The hd values are shifted in time. 
Only valid values are shifted.
}

