% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{shiftLinRandom}
\alias{shiftLinRandom}
\alias{shiftLinRandom,Positrack-method}
\alias{shiftLinRandom,ANY,ANY-method}
\title{Shift the linear position data of the animal of a random amount larger than the value of minShiftMs}
\usage{
shiftLinRandom(pt)

\S4method{shiftLinRandom}{Positrack}(pt)
}
\arguments{
\item{pt}{Positrack object}
}
\value{
Positrack object with shifted linear position (lin) values
}
\description{
This is used to do shuffling. The lin values are shifted in time. 
Only valid values are shifted.
}
