% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{shiftPositionRandom}
\alias{shiftPositionRandom}
\alias{shiftPositionRandom,Positrack-method}
\alias{shiftPositionRandom,ANY,ANY-method}
\title{Shift the position of the animal of a random amount larger than the value of minShiftMs}
\usage{
shiftPositionRandom(pt)

\S4method{shiftPositionRandom}{Positrack}(pt)
}
\arguments{
\item{pt}{Positrack object}
}
\value{
Positrack object with shifted x and y values
}
\description{
This is used to do shuffling. The position values (x and y) are shifted in time. 
Only valid values are shifted, so if you want to only shift the position only in 
one environment or a specific part of the track, first call setInvalidOutsideInterval
}
