% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{shiftPositionVector}
\alias{shiftPositionVector}
\title{Shift a the values of a vector by a random amount that is at least as large as the argument minMvMs}
\usage{
shiftPositionVector(x, timePerSampleRes, minMvMs, samplingRate)
}
\arguments{
\item{x}{A vector}

\item{timePerSampleRes}{Time in sample values (from the .dat files) between the position sample}

\item{minMvMs}{Minimum shift in ms}

\item{samplingRate}{Sampling rate of the .dat files.}
}
\description{
Shift a the values of a vector by a random amount that is at least as large as the argument minMvMs
}
\examples{
shiftPositionVector(x=1:100, timePerSampleRes=400, minMvMs = 1000, samplingRate=20000)
}

