% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{shiftPositionVectors}
\alias{shiftPositionVectors}
\title{Shift a the values of two vectors by a random amount that is at least as large as the argument minMvMs}
\usage{
shiftPositionVectors(x, y, timePerSampleRes, minMvMs, samplingRate)
}
\arguments{
\item{x}{A vector}

\item{y}{A second vector}

\item{timePerSampleRes}{Time in sample values (from the .dat files) between the position sample}

\item{minMvMs}{Minimum shift in ms}

\item{samplingRate}{Sampling rate of the .dat files.}
}
\description{
The two vectors are shifted by the same amount.
}
\examples{
shiftPositionVectors(x=1:100,y=201:300, timePerSampleRes=400, minMvMs = 1000, samplingRate=20000)
}

