% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{shiftSpeedRandom}
\alias{shiftSpeedRandom}
\alias{shiftSpeedRandom,Positrack-method}
\alias{shiftSpeedRandom,ANY,ANY-method}
\title{Shift the speed of the animal of a random amount larger than the value of minShiftMs}
\usage{
shiftSpeedRandom(pt)

\S4method{shiftSpeedRandom}{Positrack}(pt)
}
\arguments{
\item{pt}{Positrack object}
}
\value{
Positrack object with shifted speed values
}
\description{
This is used to do shuffling. The speed values are shifted in time. 
Only valid values are shifted.
}
