% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{simulateRawTrace}
\alias{simulateRawTrace}
\title{Generate raw traces with some background gaussian noise and surimposed spikes}
\usage{
simulateRawTrace(samplingRate = 20000, durationSec = 1, noiseSD = 100,
  noiseMean = 0, waveformAmplitude = 700, nClusters = 3, nChannels = 4,
  waveformDifferentiationSD = 200, minChannelScalling = 0.3,
  spikeJitterAcrossChannelMs = 0.1, maxSpikes = 10000)
}
\arguments{
\item{samplingRate}{Sampling rate of the trace}

\item{durationSec}{Total duration in second of the trace}

\item{noiseSD}{Standard deviation of gaussian noise}

\item{noiseMean}{Mean of noise}

\item{waveformAmplitude}{Negative amplitude of the generic spike waveform}

\item{nClusters}{Number of different waveforms (neurons) in the trace}

\item{nChannels}{Number of channels, 4 in case of tetrodes}

\item{waveformDifferentiationSD}{Differentiation of the waveforms of different cluster 
(gaussian noise added in generic waveform)}

\item{minChannelScalling}{Minimal value (between 0 and 1) that can be used for creating the tetorde effect 
(scalling of waveform on different channels)}

\item{spikeJitterAcrossChannelMs}{Time variation between the waveforms of a spike on the different channels.}

\item{maxSpikes}{Maximum number of spikes to include in the traces}
}
\value{
list containing trace, spikeTimes and cluId
}
\description{
Several different waveforms can be used. They are all generated from the same generic spike by adding gaussian noise to it.
}

