% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{smoothGaussian}
\alias{smoothGaussian}
\title{Smooth the values in a numeric vector using a Gaussian kernel}
\usage{
smoothGaussian(x, sd = 2, invalid = -1, type = "linear")
}
\arguments{
\item{x}{Numeric vector}

\item{sd}{The standard deviation of the Gaussian kernel used to smooth}

\item{invalid}{Numeric indicating which value should be treated as NA, by default -1.0.
The value should be a numeric.}

\item{type}{character vector indicating the type of data to smooth 
Valid values are linear, circular, degrees.
circular assumes that the vector is, i.e. the first and last values are adjacent.
degrees assumes that the vector contains degrees (0=360)}
}
\description{
The values at -1.0 are consider invalid by defaults and are not used or changed. 
Set the value of the argument degrees to TRUE if the data are circular. 
For example, if the first and last data of the vector should be considered next to each other.
}
\examples{
smoothGaussian(x=c(1:10,9:1),sd=2,invalid=-1.0,degrees=FALSE)
}
