% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{smoothhd}
\alias{smoothhd}
\alias{smoothhd,ANY,ANY-method}
\alias{smoothhd,Positrack-method}
\title{Apply some smoothing to the hd array}
\usage{
smoothhd(pt, sd)

\S4method{smoothhd}{Positrack}(pt, sd = 2)
}
\arguments{
\item{pt}{Positrack object}

\item{sd}{Standard deviation of the smoothing kernel. The unit is position samples.}
}
\value{
Positrack object with slot hd smoothed.
}
\description{
A Gaussian kernel is used for smoothing
This function is not working at the moment. A warning message will appear
The c function needs to be written
}

