% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{smoothxy}
\alias{smoothxy}
\alias{smoothxy,Positrack-method}
\alias{smoothxy,ANY,ANY-method}
\title{Apply some smoothing to the x and y position data}
\usage{
smoothxy(pt, sd)

\S4method{smoothxy}{Positrack}(pt, sd = 2)
}
\arguments{
\item{pt}{Positrack object}

\item{sd}{Standard deviation of the smoothing kernel. The unit is position samples.}
}
\value{
Positrack object with slots x and y smoothed.
}
\description{
A Gaussian kernel is used for smoothing
}
