% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\name{sortRecSessionListChronologically}
\alias{sortRecSessionListChronologically}
\title{Sort a list of RecSession objects according to the recording date of the sessions}
\usage{
sortRecSessionListChronologically(rsl)
}
\arguments{
\item{rsl}{List containing recording sessions}
}
\value{
List containing the recording sessions ordered according to their recording date.
}
\description{
Use the date in the name of the RecSession object
}
