% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties1d.R
\docType{class}
\name{SpatialProperties1d-class}
\alias{SpatialProperties1d-class}
\alias{SpatialProperties1d}
\title{A S4 class to analyze spatial properties in a 1D environment}
\description{
Use to get firing rate histogram of 1D environment (e.g. linear track).
You can get information scores, sparsity, etc.
}
\section{Slots}{

\describe{
\item{\code{session}}{Name of the recording session}

\item{\code{cmPerBin}}{Number of cm in each bin of a linear firing rate map, 2 by default}

\item{\code{smoothOccupancySd}}{Standard deviation in cm of the gaussian kernel used to smooth the occupancy map, 3 by default}

\item{\code{smoothRateHistoSd}}{Standard deviation in cm of the gaussian kernel used to smooth the firing rate histograms, 3 by default}

\item{\code{nBinRateHisto}}{Number of bins in the firing rate histograms, set according to the max position value.}

\item{\code{xSpikes}}{Position of the animal at each spike time}

\item{\code{rateHisto}}{Firing rate histogram of all neurons, stored in an array}

\item{\code{occupancy}}{Occupancy histogram}

\item{\code{cellList}}{Cell list to analyze}

\item{\code{reduceSize}}{Logical, if set to TRUE, position data will be shifted so that the minimum has a value of 0, TRUE by default}

\item{\code{peakRate}}{Vector containing the peak firing rates from the firing rate histograms}

\item{\code{infoScore}}{Information scores of the firing rate histograms}

\item{\code{sparsity}}{Sparsity of the firing rate histograms}

\item{\code{nShufflings}}{Number of shufflings to get chance levels, by default 100}

\item{\code{minShiftMs}}{Minimum time shift of the position data when doing shuffling}

\item{\code{peakRateShuffle}}{peak firing rate in the shuffling analysis}

\item{\code{infoScoreShuffle}}{Information score from the shuffling analysis}

\item{\code{sparsityShuffle}}{Sparsity from the shuffling analysis}
}}

