% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{class}
\name{SpatialProperties2d-class}
\alias{SpatialProperties2d}
\alias{SpatialProperties2d-class}
\title{A S4 class to analyze spatial properties in 2D environments}
\description{
Use to get firing rate maps of neurons in an open field.
It can get information scores, sparsity, etc.
You can use it to analyze the firing properties of grid cells or border cells.
Can do some shuffling to get the spatial properties that you would get by chance.
}
\section{Slots}{

\describe{
\item{\code{session}}{Name of the recording session}

\item{\code{cmPerBin}}{Number of cm in each bin of a firing rate map, 2 by default}

\item{\code{smoothOccupancySd}}{Standard deviation in cm of the gaussian kernel used to smooth the occupancy map, 3 by default}

\item{\code{smoothRateMapSd}}{Standard deviation in cm of the gaussian kernel used to smooth the firing rate histograms, 3 by default}

\item{\code{nColMap}}{Number of columns in the firing rate maps}

\item{\code{nRowMap}}{Number of rows in the firing rate maps}

\item{\code{nColAuto}}{Number of columns in the spatial autocorrelation}

\item{\code{nRowAuto}}{Number of rows in the spatial autocorrelation}

\item{\code{xSpikes}}{x position of the animal for each spike time}

\item{\code{ySpikes}}{y position of the animal for each spike time}

\item{\code{maps}}{Array containing the firing rate maps of the neurons}

\item{\code{occupancy}}{Matrix containing the occupancy map}

\item{\code{autos}}{Array containing the spatial autocorrelation of the firing rate maps}

\item{\code{autosDetect}}{Array with the spatial autocorrelation once the firing fields have been removed}

\item{\code{autosDoughnut}}{Array with the spatial autocorrelation containing only the ring of the 6 closest fields (excluding central field)}

\item{\code{autosDoughnutRotate}}{Array with the rotated spatial autocorrelations}

\item{\code{cellList}}{List of cells}

\item{\code{reduceSize}}{Logical indicating if the size of map should be minimized by setting smallest x and y to 0}

\item{\code{minValidBinsAuto}}{Number of valid bins to calculate a r value in the spatial autocorrelation maps}

\item{\code{peakRate}}{Peak firing rates in the maps}

\item{\code{infoScore}}{Information score of the firing rate maps}

\item{\code{sparsity}}{Sparsity of the firing rate maps}

\item{\code{borderScore}}{Border score for each firing rate maps}

\item{\code{borderCM}}{Value of CM when calculating the border score}

\item{\code{borderDM}}{Value of DM when calculating the border score}

\item{\code{borderNumFieldsDetected}}{Number of detected fields when calculating the border score}

\item{\code{borderPercentageThresholdField}}{Threshold for a bin being part of a field, expressed as percentage of peak}

\item{\code{borderMinBinsInField}}{Minimum number of bins to be considered a field}

\item{\code{mapPolarity}}{How polarized is the firing rate of the cell in the map. Each bin of the map is given an angle and firing rate.
mapPolarity is the resultant mean vector length}

\item{\code{gridScore}}{Grid score calculated form the spatial autocorrelation}

\item{\code{gridOrientation}}{Grid orientation}

\item{\code{gridSpacing}}{Grid spacing in cm}

\item{\code{gridScoreNumberFieldsToDetect}}{Maximal number of fields that will be detected}

\item{\code{gridScoreMinNumBinsPerField}}{Minimum number of bins to be considered a field}

\item{\code{gridScoreFieldThreshold}}{Minimum value of firing fields in the spatial autocorrelation map}

\item{\code{nAutoRotations}}{Number of rotations when rotating the spatial autocorrelation map}

\item{\code{AutoRotationDegree}}{Angle of rotation of the spatial autocorrelation map}

\item{\code{speedScore}}{Speed score of the neurons}

\item{\code{speedRateSlope}}{Slope of the linear regression line between speed and rate}

\item{\code{speedRateIntercept}}{Rate intercept of the linear regression line between speed and rate}

\item{\code{nShufflings}}{Number of shufflings to get chance levels, by default 100}

\item{\code{minShiftMs}}{Minimum time shift of the position data when doing shuffling}

\item{\code{peakRateShuffle}}{peak firing rate in the shuffling analysis}

\item{\code{infoScoreShuffle}}{Information score from the shuffling analysis}

\item{\code{sparsityShuffle}}{Sparsity from the shuffling analysis}

\item{\code{borderScoreShuffle}}{Border score form the shuffling analysis}

\item{\code{borderCMShuffle}}{Border CM from the shuffling analysis}

\item{\code{borderDMShuffle}}{Border DM from the shuffling analysis}

\item{\code{borderNumFieldsDetectedShuffle}}{Number of detected fields when calculating shufflings of the border score}

\item{\code{mapPolarityShuffle}}{Map polarity from the shuffling analysis}

\item{\code{gridScoreShuffle}}{Grid score from the shuffling analysis}

\item{\code{speedScoreShuffle}}{Speed score from the shuffling analysis}
}}

