% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{speedRateTuningCurve}
\alias{speedRateTuningCurve}
\alias{speedRateTuningCurve,SpatialProperties2d-method}
\alias{speedRateTuningCurve,ANY,ANY-method}
\title{Speed-rate tuning curve from spike train and positrack}
\usage{
speedRateTuningCurve(sp, st, pt, minSpeed, maxSpeed)

\S4method{speedRateTuningCurve}{SpatialProperties2d}(sp, st, pt, minSpeed = 2,
  maxSpeed = 100)
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{st}{SpikeTrain object with ifr}

\item{pt}{Positrack object}

\item{minSpeed}{Minimal speed to be considered}

\item{maxSpeed}{Maximal speed to be considered}
}
\value{
data.frame with the speed-rate tuning curves
}
\description{
Uses the ifr of cells to calculate the mean firing rate of neurons 
for different speed bands. You need to call ifr() on the st object first.
}
