% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{speedScore}
\alias{speedScore}
\alias{speedScore,ANY,ANY-method}
\alias{speedScore,SpatialProperties2d-method}
\title{Speed scores from spike train and positrack}
\usage{
speedScore(sp, st, pt, minSpeed, maxSpeed, runLm = F)

\S4method{speedScore}{SpatialProperties2d}(sp, st, pt, minSpeed = 2,
  maxSpeed = 100, runLm = F)
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{st}{SpikeTrain object with ifr}

\item{pt}{Positrack object}

\item{minSpeed}{Minimal speed to be considered}

\item{maxSpeed}{Maximal speed to be considered}

\item{runLm}{Logical, if TRUE a linear model will be build and slope and intercept calculated}
}
\value{
SpatialProperties2d object with the speed scores in the slots SpeedScore. 
If runLm argument was TRUE, then speedRateSlope and peedRateIntercept will also be filled
}
\description{
Speed scores from spike train and positrack
}

