% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{speedScoreShuffle}
\alias{speedScoreShuffle}
\alias{speedScoreShuffle,SpatialProperties2d-method}
\alias{speedScoreShuffle,ANY,ANY-method}
\title{Random speed scores from spike train and positrack}
\usage{
speedScoreShuffle(sp, st, pt, minSpeed, maxSpeed)

\S4method{speedScoreShuffle}{SpatialProperties2d}(sp, st, pt, minSpeed = 3,
  maxSpeed = 100)
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{st}{SpikeTrain object with ifr}

\item{pt}{Positrack object}

\item{minSpeed}{Minimal speed to be considered}

\item{maxSpeed}{Maximal speed to be considered}
}
\value{
SpatialProperties2d object with the random speed scores in the slots SpeedScoreShuffle
}
\description{
Random speed scores from spike train and positrack
}
