% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{spikeDetectionAccuracy}
\alias{spikeDetectionAccuracy}
\title{Get the results of spike detection from simulated data.}
\usage{
spikeDetectionAccuracy(sTimeD, sTimeT, maxJitter = 2)
}
\arguments{
\item{sTimeD}{Spike times of detected spikes}

\item{sTimeT}{Spike times of simulated spikes}

\item{maxJitter}{Max jitter to considered simulated and detected spike as the same spikes}
}
\description{
Get the results of spike detection from simulated data.
}

