% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{spikeDistanceMetric}
\alias{spikeDistanceMetric}
\alias{spikeDistanceMetric,ANY,ANY-method}
\alias{spikeDistanceMetric,SpatialProperties2d-method}
\title{Get the spike distance metric of the spikes}
\usage{
spikeDistanceMetric(sp, st, pt, startIntervals, endIntervals,
  percentileField = 75)

\S4method{spikeDistanceMetric}{SpatialProperties2d}(sp, st, pt, startIntervals,
  endIntervals, percentileField = 75)
}
\arguments{
\item{sp}{SpatialProperties2d object with valid firing rate maps}

\item{st}{SpikeTime object}

\item{pt}{Positrack object}

\item{startIntervals}{Numeric with time in sample values for time 0}

\item{endIntervals}{Numeric with time in sample values for end of intervals}

\item{percentileField}{Numeric with the percentile of bins in the maps to act as a threshold to be part of a field}
}
\value{
Matrix with clu.id, interval.no, time.sec, distance.from.field.center
}
\description{
Score based on Hardcastle et al. 2015 Neuron
}

