% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{spikeExtractionSession}
\alias{spikeExtractionSession}
\title{Perform the spike extraction for a recording session}
\usage{
spikeExtractionSession(rs, minPassHz = 800, maxPassHz = 5000,
  powerWindowSizeMs = 0.4, powerWindowSlideMs = 0.1, SDThreshold = 2,
  simultaneousSpikeMaxJitterMs = 0.2, spikeDetectionRefractoryMs = 0.5,
  waveformWindowSizeMs = 1, firstSample = 0, lastSample = -1)
}
\arguments{
\item{rs}{RecSession object}

\item{minPassHz}{Minimal frequency in Hz that is kept before calculating power}

\item{maxPassHz}{Maximal frequency in Hz that is kept}

\item{powerWindowSizeMs}{Size in ms of the sliding window to calculate power}

\item{powerWindowSlideMs}{Shift of the power window between calculation of power}

\item{SDThreshold}{Power threshold in standard deviation above the mean for detection of a spike}

\item{simultaneousSpikeMaxJitterMs}{Use to join spikes detected on several channels}

\item{spikeDetectionRefractoryMs}{Period of refractory in the detection of spikes on a single channel}

\item{waveformWindowSizeMs}{Window size used when extracting the spike waveforms.}

\item{firstSample}{First sample to consider in spike detection, by default 0, indices start at 0}

\item{lastSample}{Last sample to consider in spike detection, if not set by user, all samples will be used}
}
\description{
This function calls spikeExtractionTetrode for each tetrode of the recording session.
}
\details{
A RecSession object is used to get the information about the assignation of the recorded channels to tetrodes and the file names.
}

