% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeWaveform.R
\name{spikeGeoFeatureOneSpike}
\alias{spikeGeoFeatureOneSpike}
\title{Get the geometrical features of a single waveform}
\usage{
spikeGeoFeatureOneSpike(wf, timePoints, baselineProportion = 0.15,
  thresholdAmplitudeProportion = 0.5)
}
\arguments{
\item{wf}{Numeric vector with waveform}

\item{timePoints}{Numeric vector with time points}

\item{baselineProportion}{Numeric indicating the percentage of wf}

\item{thresholdAmplitudeProportion}{that is used to calculate the baseline}
}
\value{
spike geometrical features (baseline,amplitude,spikeDuration, firstHalfSpikeDuration, secondHalfSpikeDuration,peakAmplitudeAssymetry)
}
\description{
Get the geometrical features of a single waveform
}

