% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{spikeOnPath}
\alias{spikeOnPath}
\alias{spikeOnPath,ANY,ANY-method}
\alias{spikeOnPath,SpatialProperties2d-method}
\title{Return the data to plot spikes on the path of the animal}
\usage{
spikeOnPath(sp, st, pt)

\S4method{spikeOnPath}{SpatialProperties2d}(sp, st, pt)
}
\arguments{
\item{sp}{SpatialProperties2d}

\item{st}{SpikeTrain object}

\item{pt}{Positrack object}
}
\value{
List containing xPath, yPath, xSpike, ySpike, cluSpike
}
\description{
Return the data to plot spikes on the path of the animal
}

