% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsSpatial.R
\name{spikeOnPathPlot}
\alias{spikeOnPathPlot}
\title{Plot a spikes on the animal path}
\usage{
spikeOnPathPlot(sop, clu, name = "", outma = c(2, 2, 2, 2),
  margin = c(1.5, 1.5, 1, 1), cex.name = 0.6, cex.point = 0.5,
  plotxlim = c(0, 80), plotylim = c(0, 80), mgp.x = c(0.5, 0.05, 0),
  mgp.y = c(0.8, 0.3, 0.2), axis.x.pos = 0, axis.y.pos = 0,
  plot.axis = TRUE, xlab = "", ylab = "")
}
\arguments{
\item{sop}{List containing the data to plot the spike on path, you can get this list with the function spikeOnPath
It contains 4 objects: xSpike, ySpike, cluSpike, xPath and yPath}

\item{clu}{Clu number of the cell of interest.}

\item{name}{String with the name of the cell}

\item{outma}{Outer margins of the figure}

\item{margin}{Inner margins of the figure}

\item{cex.name}{Size of the font use for the name of the map}

\item{cex.point}{Size of the spike points}

\item{plotxlim}{Limits of the x axis}

\item{plotylim}{Limits of the y axis}

\item{mgp.x}{mgp for the x axis}

\item{mgp.y}{mgp for the y axis}

\item{axis.x.pos}{Y position of the x axis}

\item{axis.y.pos}{X position of the y axis}

\item{plot.axis}{Whether to plot the axes or not (TRUE or FALSE)}

\item{xlab}{Name to display under the x axis}

\item{ylab}{Name to display at the left of the y axis}
}
\description{
Plot the animal path from a Positrack object and the spikes in space from a SpatialProperties2d object
}

