% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{spikePCA}
\alias{spikePCA}
\title{Do PCA analysis on the spike waveforms, treating each channel independently
 
 Do PCA analysis from spike waveform, keep the first 3 or 4 principal components
 Combine the first components of each wire.}
\usage{
spikePCA(swf)
}
\arguments{
\item{swf}{Array with the spike waveforms [spike,time,channel]}
}
\value{
Matrix with the different spike features, one spike per row
}
\description{
Do PCA analysis on the spike waveforms, treating each channel independently
 
 Do PCA analysis from spike waveform, keep the first 3 or 4 principal components
 Combine the first components of each wire.
}

