% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{spikeTimeAutocorrelation}
\alias{spikeTimeAutocorrelation}
\alias{spikeTimeAutocorrelation,ANY,ANY-method}
\alias{spikeTimeAutocorrelation,SpikeTrain-method}
\title{Calculate the spike-time autocorrelation}
\usage{
spikeTimeAutocorrelation(st, binSizeMs, windowSizeMs, probability)

\S4method{spikeTimeAutocorrelation}{SpikeTrain}(st, binSizeMs = 1,
  windowSizeMs = 200, probability = FALSE)
}
\arguments{
\item{st}{SpikeTrain object}

\item{binSizeMs}{Default is 1}

\item{windowSizeMs}{Default is 200. This means that autocorrelation ranges from -windowSizeMs to windowSizeMs}

\item{probability}{If TRUE, will calculate the probability of a spike in a given bin instead of the spike count}
}
\value{
SpikeTrain object with autocorrelation in slot auto
}
\description{
Each spike is treated in turn as a reference spike.
The number of spikes or probability to observe a spike around the reference spike is calculated.
You can set the bins size in ms and and the time window for which you want to do the analysis on.
The results are saved in the slot auto and autoTimePoints.
}

