% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{spikeTimeAutocorrelationAsDataFrame}
\alias{spikeTimeAutocorrelationAsDataFrame}
\alias{spikeTimeAutocorrelationAsDataFrame,SpikeTrain-method}
\alias{spikeTimeAutocorrelationAsDataFrame,ANY,ANY-method}
\title{Get spike-time autocorrelation as data.frame}
\usage{
spikeTimeAutocorrelationAsDataFrame(st)

\S4method{spikeTimeAutocorrelationAsDataFrame}{SpikeTrain}(st)
}
\arguments{
\item{st}{SpikeTrain object}
}
\value{
data.frame with spike-time autocorrelation
}
\description{
Get spike-time autocorrelation as data.frame
}
