% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsGeneric.R
\name{spikeTimeAutocorrelationPlot}
\alias{spikeTimeAutocorrelationPlot}
\title{Plot a single spike-time autocorrelation}
\usage{
spikeTimeAutocorrelationPlot(x, y, name = "", axis.y.pos = NA,
  axis.x.pos = 0, axis.y.las = 2, mgp.x = c(0.5, 0.1, 0.1),
  mgp.y = c(0.9, 0.2, 0.1), xlab = "Time (ms)", ylab = "Spikes",
  plotxlim = NA, plotylim = NA, outma = c(0.5, 0.5, 0.5, 0.5),
  margin = c(1.5, 1.5, 1, 0.3), xaxis.at = NA, yaxis.at = NA,
  add.text = "", add.text.pos = c(0, 0.5), ...)
}
\arguments{
\item{x}{Numeric vectors with the x values}

\item{y}{Numeric vectors with the y values}

\item{name}{Character vectors containing the name of the graph}

\item{axis.y.pos}{Position of the y axis}

\item{axis.x.pos}{Position of the x axis}

\item{axis.y.las}{Orientation of the letters for the label of the y axis}

\item{mgp.x}{mgp for x axis}

\item{mgp.y}{mgp for y axis}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{plotxlim}{Limit of the x axis}

\item{plotylim}{Limit of the y axis}

\item{outma}{outma for setting par}

\item{margin}{margin for setting par}

\item{xaxis.at}{Where the tics are shown on the x axis}

\item{yaxis.at}{Where the tics are shown on the y axis}

\item{add.text}{Text to add to the plot}

\item{add.text.pos}{Position of the text to add, format (x,y)}

\item{...}{Passed to the lines function}
}
\description{
Plot a single spike-time autocorrelation
}

