% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsGeneric.R
\name{spikeTimeAutocorrelationsPlot}
\alias{spikeTimeAutocorrelationsPlot}
\title{Plot several spike-time autocorrelation plot on the same page}
\usage{
spikeTimeAutocorrelationsPlot(autos, timePoints, names)
}
\arguments{
\item{autos}{A matrix containing autocorrelations}

\item{timePoints}{A numerical vector with the time point for each value of the autocorrelation}

\item{names}{A character vector containing the name of each map in the array}
}
\description{
Plot several spike-time autocorrelation plot on the same page
}

