% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{spikeTimeCrosscorrelationAsDataFrame}
\alias{spikeTimeCrosscorrelationAsDataFrame}
\alias{spikeTimeCrosscorrelationAsDataFrame,SpikeTrain-method}
\alias{spikeTimeCrosscorrelationAsDataFrame,ANY,ANY-method}
\title{Get spike-time crosscorelation between cells as data.frame}
\usage{
spikeTimeCrosscorrelationAsDataFrame(st)

\S4method{spikeTimeCrosscorrelationAsDataFrame}{SpikeTrain}(st)
}
\arguments{
\item{st}{SpikeTrain object}
}
\value{
data.frame with spike-time crosscorrelation between neurons
}
\description{
Get spike-time crosscorelation between cells as data.frame
}
