% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{spikeTimeCrosscorrelationEvents}
\alias{spikeTimeCrosscorrelationEvents}
\alias{spikeTimeCrosscorrelationEvents,SpikeTrain-method}
\alias{spikeTimeCrosscorrelationEvents,ANY,ANY-method}
\title{Calculate the spike-time crosscorrelation between the spike trains and a list of events}
\usage{
spikeTimeCrosscorrelationEvents(st, binSizeMs = 1, windowSizeMs = 200,
  probability = FALSE)

\S4method{spikeTimeCrosscorrelationEvents}{SpikeTrain}(st, binSizeMs = 1,
  windowSizeMs = 200, probability = FALSE)
}
\arguments{
\item{st}{SpikeTrain object}

\item{binSizeMs}{Default is 1}

\item{windowSizeMs}{Default is 200, meaning that it ranges from + and - 200}

\item{probability}{If TRUE, will calculate the probability of a spike in a given bin instead of the spike count}
}
\value{
st SpikeTrain object with the slot crossEvents filled
}
\description{
Each event is treated in turn as a reference event.
The number of spikes or probability to observe a spike around the reference event is calculated.
You can set the bins size in ms and and the time window for which you want to do the analysis on.
}
