% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{spikeTimeCrosscorrelationEventsAsDataFrame}
\alias{spikeTimeCrosscorrelationEventsAsDataFrame}
\alias{spikeTimeCrosscorrelationEventsAsDataFrame,ANY,ANY-method}
\alias{spikeTimeCrosscorrelationEventsAsDataFrame,SpikeTrain-method}
\title{Get spike-time crosscorelation to events as data.frame}
\usage{
spikeTimeCrosscorrelationEventsAsDataFrame(st)

\S4method{spikeTimeCrosscorrelationEventsAsDataFrame}{SpikeTrain}(st)
}
\arguments{
\item{st}{SpikeTrain object}
}
\value{
data.frame with spike-time crosscorrelation to events
}
\description{
Get spike-time crosscorelation to events as data.frame
}

