% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{class}
\name{SpikeTrain-class}
\alias{SpikeTrain}
\alias{SpikeTrain-class}
\title{A S4 class to represent the spike trains of one recording session}
\description{
This class is used to analyse the spike trains of neurons. You can calculate mean firing rates,
instantaneous firing rates, spike-time autocorrelations, spike-time crosscorrelations,
crosscorrelation between spike trains and events, etc.
You can limit the analysis to some time interval by using the method setInterval.
Similarly, you can limit the analysis to some cells with the method setCellList.
}
\section{Slots}{

\describe{
\item{\code{session}}{Name of the recording session}

\item{\code{path}}{Directory where the recording session is located}

\item{\code{samplingRate}}{Sampling rate of the electrophysiological data}

\item{\code{res}}{Spike time in sample value}

\item{\code{time}}{Spike time in seconds}

\item{\code{clu}}{Cluster id for each spike}

\item{\code{nCells}}{Number of cells in the recording session}

\item{\code{nSpikes}}{Number of spikes in the recording session}

\item{\code{nSpikesPerCell}}{Number of spikes for each cell}

\item{\code{startInterval}}{Sample values of the beginning of intervals. To limit the analysis to some intervals.
Only data within the intervals are considered}

\item{\code{endInterval}}{Sample values of the end of intervals. Only data within the intervals are considered.}

\item{\code{startResIndexc}}{Index in the spike arrays for the start of intervals. Index is for a C array with 0 indexing.
The spike at the index is to be considered for analysis}

\item{\code{endResIndexc}}{Index in the spike arrays for the start of intervals. Index is for a C array with 0 indexing.
The spike at the intdex is to be considered for analysis}

\item{\code{events}}{Time in sample number for some events}

\item{\code{cellList}}{Cell list}

\item{\code{auto}}{Matrix holding spike-time autocorrelation}

\item{\code{autoMsPerBin}}{Ms per bin in spike-time autocorrelation}

\item{\code{autoTimePoints}}{Time points for data points in the spike-time autocorrelation}

\item{\code{autoProbability}}{Logical, whether the spike-time autocorrelation contains probability values or spike count}

\item{\code{autoCOM}}{Center of mass of the positive half of each spike-time autocorrelation}

\item{\code{cross}}{Matrix holding spike-time crosscorrelation between spikes and events}

\item{\code{crossMsPerBin}}{Ms per bin in spike-time crosscorrelation with events}

\item{\code{crossTimePoints}}{Time points for data points in the spike-time crosscorrelation}

\item{\code{crossProbability}}{Logical, whether the spike-time crosscorrelation to events contains probability values or spike count}

\item{\code{crossEvents}}{Matrix holding spike-time crosscorrelation between spikes and events}

\item{\code{crossEventsMsPerBin}}{Ms per bin in spike-time crosscorrelation with events}

\item{\code{crossEventsTimePoints}}{Time points for data points in the spike-time crosscorrelation to events}

\item{\code{crossEventsProbability}}{Logical, whether the spike-time crosscorrelation to events contains probability values or spike count}

\item{\code{cellPairList}}{Data frame containing pairs of cells}

\item{\code{ifrKernelSdMs}}{Standard deviation of a gaussian kernel used to calculate the instantaneous firing rate}

\item{\code{ifrWindowSizeMs}}{Window size of the instantaneous firing rate array}

\item{\code{ifrSpikeBinMs}}{Size of the bins in which the number of spikes are counted}

\item{\code{ifr}}{Matrix containing the instantaneous firing rate of the neurons}

\item{\code{ifrTime}}{Time associated with each window of the instantaneous firing rate}

\item{\code{meanFiringRate}}{Mean firing rate of the neurons}

\item{\code{isolationDistance}}{Isolation distance of each cluster}

\item{\code{refractoryRatio}}{Refractory ratio of each cluster}

\item{\code{crossRefractoryRatio}}{Refractory ratio found in the spike-time crosscorrelation between a cluster
and all other clusters. The smallest ratio is used.}
}}

