% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{spikeTriggeredFiringRateMap2d}
\alias{spikeTriggeredFiringRateMap2d}
\alias{spikeTriggeredFiringRateMap2d,SpatialProperties2d-method}
\alias{spikeTriggeredFiringRateMap2d,ANY,ANY-method}
\title{Calculate the spike-triggered firing rate maps of neurons using a SpikeTrain and Positrack objects}
\usage{
spikeTriggeredFiringRateMap2d(sp, st, pt, minIsiMs, maxIsiMs)

\S4method{spikeTriggeredFiringRateMap2d}{SpatialProperties2d}(sp, st, pt,
  minIsiMs, maxIsiMs)
}
\arguments{
\item{sp}{SpatialProperties1d object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack object}

\item{minIsiMs}{Minimal interspike interval to consider in ms}

\item{maxIsiMs}{Maximal interspike interval to consider in ms}
}
\value{
SpatialProperties2d object with the spike-triggered firing rate maps
}
\description{
Each spike is treated as a reference spike in turn. The map is constructed from the data
following the reference spikes by shifting the x and y coordinate so that the position of the 
agent at the time of the reference spike is 0,0.
}
\details{
The occupancy map and the firing rate maps are smoothed with a Gaussian kernel
The amount of smoothing is determined by slots smoothOccupancySd and smoothRateMapSd of sp

You can set the temporal limit for the data used to construct the map with minIsiMs and maxIsiMs
}
