% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeadDirection.R
\docType{methods}
\name{spikeTriggeredHeadDirectionHisto}
\alias{spikeTriggeredHeadDirectionHisto}
\alias{spikeTriggeredHeadDirectionHisto,HeadDirection-method}
\alias{spikeTriggeredHeadDirectionHisto,ANY,ANY-method}
\title{Calculate the spike-triggered head-direction histogram using a HeadDirection, SpikeTrain and Positrack objects}
\usage{
spikeTriggeredHeadDirectionHisto(hd, st, pt, minIsiMs, maxIsiMs)

\S4method{spikeTriggeredHeadDirectionHisto}{HeadDirection}(hd, st, pt, minIsiMs,
  maxIsiMs)
}
\arguments{
\item{hd}{HeadDirection object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack object}

\item{minIsiMs}{Minimal interspike interval to consider in ms}

\item{maxIsiMs}{Maximal interspike interval to consider in ms}
}
\value{
HeadDirection object with the spike-triggered rate histograms in the hist slot
}
\description{
Each spike is treated as a reference spike in turn, and set to phase 180 degree. 
The histogram is constructed from the data following the reference spikes 
by shifting the head direction data so that the head direction of the 
agent at the time of the reference spike is 180 degree.
}
\details{
The occupancy histogram and the firing rate histogram are smoothed with a Gaussian kernel.
The amount of smoothing is determined by slots smoothOccupancySd and smoothRateHistoSd of the HeadDirection object.

You can set the temporal limit for the data used to construct the histogram with minIsiMs and maxIsiMs
}
