% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeWaveform.R
\docType{class}
\name{SpikeWaveform-class}
\alias{SpikeWaveform-class}
\alias{spikeWaveform}
\title{A S4 class to represent the spike waveforms of neurons}
\description{
This class is used to get the mean waveform of a neuron. 
It can also extract some features from the spike waveform.
}
\section{Slots}{

\describe{
\item{\code{session}}{Name of the recording session}

\item{\code{path}}{Directory where the recording session is located}

\item{\code{samplingRate}}{Sampling rate of the electrophysiological data}

\item{\code{cellList}}{Cell list}

\item{\code{wf}}{Matrix holding spike waveforms of the neurons}

\item{\code{wfCluId}}{Character vector indicating which waveforms belong to which neuron}

\item{\code{wfMsPerBin}}{Ms per bin in spike-time autocorrelation}

\item{\code{wfTimePoints}}{Time points for data points in the spike-time autocorrelation}
}}

