% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{spikeWaveformFromTraces}
\alias{spikeWaveformFromTraces}
\title{Create the spike waveform array
   
 Returns a 3D array with the waveforms of each spike for each channel.
 array[spike,time,channel]}
\usage{
spikeWaveformFromTraces(traces, res, window = 20)
}
\arguments{
\item{traces}{Matrix with the traces for each channel. Each column is a channel}

\item{res}{Numeric with the time stamps of the spikes in sample value}

\item{window}{Number of samples for each spike}
}
\value{
array[spike,time,channel]
}
\description{
Create the spike waveform array
   
 Returns a 3D array with the waveforms of each spike for each channel.
 array[spike,time,channel]
}
