% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{statsAsDataFrame}
\alias{statsAsDataFrame}
\alias{statsAsDataFrame,SpatialProperties2d-method}
\alias{statsAsDataFrame,ANY,ANY-method}
\title{Return firing rate map statisitc as a data.frame}
\usage{
statsAsDataFrame(sp, shuffle)

\S4method{statsAsDataFrame}{SpatialProperties2d}(sp, shuffle = FALSE)
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{shuffle}{Logical, if TRUE returns the stats of the shuffling analysis}
}
\value{
data.frame with the firing rate map stats.
}
\description{
Return firing rate map statisitc as a data.frame
}
