% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinIntervals.R
\name{timeWithinIntervals}
\alias{timeWithinIntervals}
\title{Get a logical vector to know if time points are within a series of intervals}
\usage{
timeWithinIntervals(x, s, e)
}
\arguments{
\item{x}{Integer vector representing time points}

\item{s}{Integer vector representing the start of intervals}

\item{e}{Integer vector representing the end of intervals}
}
\value{
Logical, whether the time points are within the intervals
}
\description{
Only time points that are within the limit of the interval will return TRUE.
For example time point 2 is not in the interval 2 4.
Only 3 would return TRUE in this case.
}

