% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeWaveform.R
\docType{methods}
\name{waveformCharacteristics}
\alias{waveformCharacteristics}
\alias{waveformCharacteristics,ANY,ANY-method}
\alias{waveformCharacteristics,SpikeWaveform-method}
\title{Calculate waveform characteristics from the mean waveform of each cluster.}
\usage{
waveformCharacteristics(sw)

\S4method{waveformCharacteristics}{SpikeWaveform}(sw)
}
\arguments{
\item{sw}{SpikeWaveform object}
}
\value{
data.frame with the different waveform features for each neuron
}
\description{
You first need to run meanWavefrom() before calling this function.
The wire with the largest range is used.
The characteristics that are calculated are amplitude, duration,
duration before trough, duration after trough,
peak from baseline before trough,
peak from baseline after trough,
peak amplitude assymetry
}

