% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whdFromPositrack.R
\name{whdAlignedTtlPositrack}
\alias{whdAlignedTtlPositrack}
\title{Try to realign the up part of ttl pulses from .dat file and the positrack frames for a single trial.}
\usage{
whdAlignedTtlPositrack(up, posi)
}
\arguments{
\item{up}{Time of up phase of ttl pulses}

\item{posi}{Data from the positrack file}
}
\description{
One assumption that is made is that there is variability in the intervals between frames in the positrack software
These time intervals can be measured from the positrack file (startProcTime) and the ttl pulse
When there is no problem of alignment, the correlation between intervals of ttl and positrack is above .97
If there is an additional ttl detected, then the correlation goes down. 
How low it is depends at which point the alignment brake
}
