% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{writeFetFile}
\alias{writeFetFile}
\title{Write the fet file}
\usage{
writeFetFile(rs, res, fet, tetrodeNumber)
}
\arguments{
\item{rs}{RecSession object}

\item{res}{Numeric with the time stamps of the spikes in sample value}

\item{fet}{Matrix with spike features, one spike per row}

\item{tetrodeNumber}{Tetrode number}
}
\description{
Write the fet file
}

