% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{writeMmFile}
\alias{writeMmFile}
\title{Write the mm file for each tetrode}
\usage{
writeMmFile(rs, fet, tetrodeNumber)
}
\arguments{
\item{rs}{RecSession object}

\item{fet}{Matrix with spike features, one spike per row}

\item{tetrodeNumber}{Tetrode number}
}
\description{
This is a legacy file created by Csicsvari detection programs
 relectro doesn't use it but it is useful if manual clustering is done with sgclust5b
 
 The content of the file is the mean and std of each feature. 
 Need to check if this is correct
}
