% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{writeParTetrodeFile}
\alias{writeParTetrodeFile}
\title{Write the par file for each tetrode}
\usage{
writeParTetrodeFile(rs, res, fet, tetrodeNumber)
}
\arguments{
\item{rs}{RecSession object}

\item{res}{Numeric with the time stamps of the spikes in sample value}

\item{fet}{Matrix with spike features, one spike per row}

\item{tetrodeNumber}{Tetrode number}
}
\description{
This is a legacy file created by Csicsvari detection programs
 relectro don't use it but it is usefull if manual clustering is done with kluster or sgclust5b
}
\details{
Here is an example            
16 4 50     # total number of electrodes, number of electrodes for this group,  sampling interval (in microseconds)
0 1 2 3     # electrode IDs
10 2        # refractory sample index after detection, RMS integration window length
90          # approximate firing frequency in Hz
16 8        # number of samples in each waveform, sample index of the peak
12 6        # window length to realign the spikes, sample index of the peak (detection program)
4 4         # number of samples (before and after the peak) to use for reconstruction and features
3 16        # number of principal components (features) per electrode, number of samples used for the PCA
800.        # high pass filter frequency (in Hz)
}
