% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{centerOfMass}
\alias{centerOfMass}
\title{Calculate the center of mass of a matrix}
\usage{
centerOfMass(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
Numeric of length 2 with the x and y coordinate of the center fo mass
}
\description{
Calculate the center of mass of a matrix
}

