% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clusterIsolationCheck}
\alias{clusterIsolationCheck}
\title{Check if clusters of a recording session are well isolated from one another}
\usage{
clusterIsolationCheck(rs)
}
\arguments{
\item{rs}{RecSession}
}
\value{
A list including a data.frame.
The check element is a logical indicating which clusters are likely to be poorly isolated.
}
\description{
This function returns a data.frame to assess the quality of the clusters.
For each neuron, you get an isolation distance, a refractory period ratio of the spike-time autocorrelation,
and a refractory period ratio of the spike-time crosscorrelations involving this neuron.
}
\details{
The data.frame is returned within a list so that the function can be used with runOnSessionList
}
