% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{createSpkFile}
\alias{createSpkFile}
\title{Create the spk file for each tetrode
   
 Use the traces and spike times given as arguments}
\usage{
createSpkFile(traces, rs, res, window = 32, tetrodeNumber, append = 0)
}
\arguments{
\item{traces}{Matrix with the electrophysiological traces}

\item{rs}{RecSession}

\item{res}{Numeric with the time stamps of the spikes in sample value}

\item{window}{Number of ms that will be shown}

\item{tetrodeNumber}{Tetrode number used in the file name}

\item{append}{If set to 1, will append to the file.}
}
\description{
Create the spk file for each tetrode
   
 Use the traces and spike times given as arguments
}

