% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeExtraction.R
\name{detectSpikesFromTrace}
\alias{detectSpikesFromTrace}
\title{Detect spikes times in a filtered signal.}
\usage{
detectSpikesFromTrace(data, samplingRate, powerWindowSizeMs, powerWindowSlideMs,
  SDThreshold, spikeDetectionRefractoryMs)
}
\arguments{
\item{data}{Vector containing the spikes and some noise}

\item{samplingRate}{Sampling rate of the trace}

\item{powerWindowSizeMs}{Window size when calculating power (root mean square)}

\item{powerWindowSlideMs}{Shift of the window in ms between estimation of power}

\item{SDThreshold}{Power threshold for spike detection.}

\item{spikeDetectionRefractoryMs}{Refractory period in the spike detection}
}
\value{
list containing rms, rmsT, rmsSD, rmsMean, rmsThreshold, spikeTime, spikePower, spikeTrough
}
\description{
This function works on the data of a single channel. 
It calculates the power with the function powerRootMeanSquare(), and calculates the mean and standard deviation.
It calls identifySpikeTimes() to get the spike times and other parameters of the detected spikes.
}
\details{
If there are low frequency components in the signal, they should be filtered out.
}

