% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsSpatial.R
\name{firingRateMapsPlot}
\alias{firingRateMapsPlot}
\title{Plot several firing rate maps on the same page}
\usage{
firingRateMapsPlot(maps, names, ncol = 5, nrow = 6)
}
\arguments{
\item{maps}{A 3d array containing maps (x,y,clu)}

\item{names}{A character vector containing the name of each map in the array}

\item{ncol}{Number of columns of plots per page}

\item{nrow}{Number of rows of plots per page}
}
\description{
Plot 2-dimensional representation of firing rate for several maps
This is not currently being developed.
}

