% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{firingRateMapsRotation}
\alias{firingRateMapsRotation}
\alias{firingRateMapsRotation,SpatialProperties2d-method}
\alias{firingRateMapsRotation,ANY,ANY-method}
\title{Rotate the firing rate maps of a SpatialProperties2d object}
\usage{
firingRateMapsRotation(sp, rotationDegrees)

\S4method{firingRateMapsRotation}{SpatialProperties2d}(sp,
  rotationDegrees = 10)
}
\arguments{
\item{sp}{SpatialProperties2d object}

\item{rotationDegrees}{Degrees of the rotation}
}
\value{
SpatialProperties2d object with the rotated maps
}
\description{
Rotate the firing rate maps of a SpatialProperties2d object
}
