% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecSession.R
\docType{methods}
\name{getIntervalsStimulation}
\alias{getIntervalsStimulation}
\alias{getIntervalsStimulation,RecSession-method}
\alias{getIntervalsStimulation,ANY,ANY-method}
\title{Get the time intervals in sample values for trials for a given stimulation}
\usage{
getIntervalsStimulation(rs, stimulation = "lt")

\S4method{getIntervalsStimulation}{RecSession}(rs, stimulation = "lt")
}
\arguments{
\item{rs}{A RecSession object}

\item{stimulation}{The name of a stimulation}
}
\value{
matrix with 2 columns containing the start and end of each trial in the stimulation
}
\description{
Get the time intervals in sample values for trials for a given stimulation
}
