% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectroProject.R
\docType{methods}
\name{getSessionList}
\alias{getSessionList}
\alias{getSessionList,ElectroProject-method}
\alias{getSessionList,ANY,ANY-method}
\title{Return a list of RecSession objects that have some common properties}
\usage{
getSessionList(ep, clustered = "", region = "", environment = "",
  stim = "", fileExtension = "")

\S4method{getSessionList}{ElectroProject}(ep, clustered = "", region = "",
  environment = "", stim = "", fileExtension = "")
}
\arguments{
\item{ep}{ElectroProject object}

\item{clustered}{logical indicating whether the session should be clustered or not}

\item{region}{Set to a given brain region to select only sessions with tetrodes in this brain region}

\item{environment}{Set to a given environment code to select only sessions during which this environment was presented}

\item{stim}{Set to a given stimulation code to select only sessions during which this stimulatoin was presented}

\item{fileExtension}{Keep sessions that have a session file ending with the value of fileExtension (e.g. kld-19021016-0101.fileExtension)}
}
\value{
list of RecSession objects
}
\description{
Return a list of RecSession objects that have some common properties
}
