% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialProperties2d.R
\docType{methods}
\name{gridSpacing}
\alias{gridSpacing}
\title{Calculate the grid spacing from the spatial autocorrelations}
\usage{
gridSpacing(sp)
}
\arguments{
\item{sp}{SpatialProperties2d object}
}
\value{
SpatialProperties2d object with the grid spacing in the slot gridSpacing
}
\description{
Calculate the grid spacing from the spatial autocorrelations
}

