% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeadDirection.R
\docType{methods}
\name{headDirectionStats}
\alias{headDirectionStats}
\alias{headDirectionStats,HeadDirection-method}
\alias{headDirectionStats,ANY,ANY-method}
\title{Calculate the head direction statistics for the head direction rate histograms with a HeadDirection, SpikeTrain and Positrack objects}
\usage{
headDirectionStats(hd, st, pt, newHisto = TRUE)

\S4method{headDirectionStats}{HeadDirection}(hd, st, pt, newHisto = TRUE)
}
\arguments{
\item{hd}{HeadDirection object}

\item{st}{SpikeTrain object}

\item{pt}{Positrack}

\item{newHisto}{logical indicating if new rate histograms are calculated when calling the function.
Default value is TRUE}
}
\value{
HeadDirection object with the statistics in the slots peakRates, vectorLength and meanDirection
}
\description{
Calculate the head direction statistics for the head direction rate histograms with a HeadDirection, SpikeTrain and Positrack objects
}
