% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotsSpatial.R
\name{linearRatePlots}
\alias{linearRatePlots}
\title{Plot several linear rate histograms from a SpatialProperties1D object}
\usage{
linearRatePlots(sp1, fn = "page.full.plot.pdf")
}
\arguments{
\item{sp1}{SpatialProperties1d object}

\item{fn}{file name}
}
\description{
In development
}

