% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{makePairs}
\alias{makePairs}
\title{Returns possible pairs from one or two vectors of values}
\usage{
makePairs(cl1 = "", cl2 = NULL)
}
\arguments{
\item{cl1}{Numeric vector containing a vector of values}

\item{cl2}{Optional argument containing a second vector of values}
}
\value{
A data.frame containing the pairs
}
\description{
If only on vector of values is given, all possible pairs within these values are returned.
If two vectors of values are given, only pairs across the two vectors are returned.
}
\examples{
makePairs(cl1=1:10)
}
