% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positrack.R
\docType{methods}
\name{speedFilter}
\alias{speedFilter}
\alias{speedFilter,ANY,ANY-method}
\alias{speedFilter,Positrack-method}
\title{Set position data for which the animal speed was not between a given range to NA}
\usage{
speedFilter(pt, minSpeed, maxSpeed)

\S4method{speedFilter}{Positrack}(pt, minSpeed = 3, maxSpeed = 100)
}
\arguments{
\item{pt}{Positrack object}

\item{minSpeed}{Minimal running speed}

\item{maxSpeed}{Maximal running speed}
}
\value{
Positrack object with invalid values outside the speed range.
}
\description{
The slots x, y, hd, speed are affected.
Please note that the speed array is affected.
}

