% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeTrain.R
\docType{methods}
\name{spikeTimeAutocorrelationCenterOfMass}
\alias{spikeTimeAutocorrelationCenterOfMass}
\alias{spikeTimeAutocorrelationCenterOfMass,ANY,ANY-method}
\alias{spikeTimeAutocorrelationCenterOfMass,SpikeTrain-method}
\title{Calculate spike-time autocorrelation center of mass}
\usage{
spikeTimeAutocorrelationCenterOfMass(st)

\S4method{spikeTimeAutocorrelationCenterOfMass}{SpikeTrain}(st)
}
\arguments{
\item{st}{SpikeTrain object}
}
\description{
Only the positive half of the spike-time autocorrelation is used.
Only positive values should be returned by this function.
}

