% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whdFromPositrack.R
\name{whdAlignedTtlPositrack}
\alias{whdAlignedTtlPositrack}
\title{Try to realign the up part of ttl pulses from .dat file and the positrack frames for a single trial.}
\usage{
whdAlignedTtlPositrack(up, posi, ccLength = 20)
}
\arguments{
\item{up}{Time of up phase of ttl pulses}

\item{posi}{Data from the positrack file}

\item{ccLength}{Lenght of the segment of data to use for crosscorrelation when a delay mismach is found}
}
\description{
One assumption that is made is that there is variability in the intervals between frames in the positrack software
These time intervals can be measured from the positrack file (startProcTime) and the ttl pulses.
When there is no problem of alignment, the correlation between intervals of ttl and positrack is above .97
If there is an additional ttl detected, then the correlation goes down.
}
\details{
The intervals between ups is calculated (dup)
The intervals between the posi$startProcTime is calculated (dposi)
We search for points where there is a large difference between dup and dposi.
The crosscorrelation function for the data points just after the large difference is calculated. 
If the peak is before or after 0, then an up or a posi is removed.
}
